Fixes to getLower/UppercaseUtf8()
On e.g. ARM and POWER char is unsigned (yeah because having it signed
wasn't stupid enough they had to make it implementation defined). Thus
the check for > 0 for ASCII characters doesn't work, instead check if
the most significant bit is set.

Also I think the size argument should take the end of the string into
account so as to never read past it even for invalid characters.