Get rid of HashMap use in StringUtils, add join
Instead of a HashMap just use an appropriately sized array on the stack,
this should be quite a bit faster (alloc on stack is often just a
pointer bump). Also 256 bytes leaves plenty of room on a default
size stack.

Also add a join function similiar to Python's string.join(). I got a bit
crazy here and it's in fact a template that works with anything
supporting operator+=

We may want to use template functions in the futue to get rid of
variants with different types too.

This also synchronizes with my ad::util variant