Add list of selected variables to JSON output
This allows matching the result columns to variables without access to the
original query or parsing the query in the response. This also makes our JSON
output a bit more like the virtuoso JSON interface allowing better code sharing
in Aqqu where the query is only accessible as a string when building the result
table. While we are here add a TODO for using a JSON library.