First try at allowing threads to wait on results
ResultTable now holds a condition_variable and accompanying mutex.  This is
used to signal the finishing of a table to threads waiting on it.  I'm not
happy with the LRUCache changes yet though, would be much nicer if we could
keep all entries strictly read-only.

Currently this hasn't been tested with multiple threads but the machinery is
already used with one thread.