Prevent copying/assigning to ResultTable + format
with ResulTable(s) being used for waiting on a result copying or assigning them
becomes difficult to handle. Thankfully we don't actually do that so delete the
copy and assignment constructors/operator=s and thus make sure it never
happens. While we are here format it and also fix some signed-unsigned
comparisons in IndexTest