Fix issue #56 by reissuing small pread()s
`pread()` has a size limit around 2.5 GB on Linux see
[here](https://stackoverflow.com/questions/36565209/pread-for-very-large-files)

While a bit weird on 64 bit systems checking the actually read number of bytes
is probably a good idea anyway. While we are here clean up `getLastOffset()`
and remove unused (except for its test) `getTrailingOffT()` which did pretty
much the same with a copy of the code but returning the read offset
directly. This also gets rid of a two more `seek()` (except for getting the
file size). More cleanup is likely needed.