Auto-detect project directory, don't use subshells
Automatically detect the project directory using ${BASH_SOURCE[0]}.

Don't use subshells for starting the ServerMain, it introduced a subtle
bug since the SERVER_PID was actually the PID of the subshell.  Turns
out there's no easy way to get the correct PID out of the subshell, so
instead we just use pushd/popd and no subshell.