Enable End-to-End test during docker build
This enables testing using our e2e/e2e.sh script during `docker build`.
The tests are executed in the final runtime stage of the multi stage
Dockerfile which makes sure that we are testing the final configuration.

This also leaves the queryit.py tool as well as the example queries in
/app/e2e_data in the final container. While these are not strictly
needed they offer useful examples.

In a later step we may replace the Travis CI build with just building
the docker container there as well. This would get rid of the version
hacks