End-to-End testing using docker
Since Travis CI only supports ancient Ubuntu versions we need a lot of
hacks to make things work. At the same time we use docker for normal
builds. So we can simplify things by just doing the end-to-end tests in
docker too. This needs a bit more flexibility in the e2e/e2e.sh script.

Also so we don't run the end-to-end tests (and their test data download)
on each docker build we introduce a docker-compose.test.yml for docker
cloud. On Travis we manually run the appropriate docker commands