Efficient Permutation creation
The temporary relation in CreatePermutationImpl is not MmapVector.
Thus it causes no trouble even if the permutation is really really
large.

The Multiplicities are efficiently created for the second order key
(O in POS) and then exchanged with the "sister" permutation.

Several changes In MmapVector and IndexMetaData subtypes for convenience