Fix broken notext queries, revert if needed
I'm not sure if we want to keep using the broken queries because they
might have been broken in the paper too, but both versions are in the
simon-queries.txt file and we can always just revert this commit.

That said we can also not be sure if they were indeed fixed for the
query but this was never commited so I would go with just fixing them to
give non-empty results.