Simplify getFollowBlockForLhs()
This is in preparation for tackling the assertion failure in #206.
Turns out fixing this function looks a bit harder as it's not clear what
to return so the rest doesn't crash. Currently the offending query
triggers the AD_CHECK(false) so we still need to figure out what to do
about it