Enforce insertion order RuntimeInformation JSON
This ensures that converting from JSON directly to YAML results in
a human readable format. Without it children are printed before
their parents which makes the YAML very hard to read.

While the JSON, as included in QLever responses, is less readable,
passing it through e.g. 'jq .' results in a readable performance digest
now too.