Separation dot without space and prefixed name
Previously with a prefixed name a final dot was not recognized as
separating triple blocks.

Sadly in SPARQL this is quite messy.

For example the following query works (in Blazegraph):

    SELECT ?item WHERE {
       ?item wdt:P31 wd:Q2934.?item wdt:P39 wd:Q41240317
    }

Using `^` reversing the following also works

    SELECT ?item WHERE {
       ?item wdt:P31 wd:Q2934. wd:Q41240317 ^wdt:P39 ?item
    }

However removing the ` ` after the `.` breaks parsing even though it's
not needed at the same position when the `?` disambiguates.

So yeah we really should use a proper parser that naturally handles this
weirdness.