Fix getVar…Columns() for TwoColumnJoin fixes #256
TwoColumnJoin currently only implements a special case. The more general
case is prevented from use by a high cost (this is the current state of
affairs not introduced by this PR).

However for that special case the column mapping was bogus. The old code
assumed that the columns are both sides concatenated with the join
columns removed from the right side. So the same as for normal Join.
However with the special case one side is only join columns and we take
the column mapping always from the other side.