Fix FILTER (lang(…) = …) for predicate paths #276
We were trying to use language tagged predicates on predicate paths.
Instead if the predicate path isn't simple we have to fallback on using
language predicates on the literal variable.

This change also means that if a predicate path mixes predicates with
literal subjects and non-literal subjects we only get the literals when
FILTER (lang(?var) = "…") is used.

The spec at

https://www.w3.org/TR/sparql11-query/#func-lang

doesn't make it clear what should happen when applied to a non-literal.
Our behaviour matches the Wikidata Query Service though.