Make cache capacity memory size based
Instead of limiting the number of elements in the cache limit the total
memory used by the cached elements. This currently uses a free template
function to let elements give their size in memory. For the results in
QLever we currently ignore the memory used by RuntimeInformation. We
also ignore the memory used by the cache itself