unique_ptr root Operation in QueryExecutionTree
The Operation is owned so it should be a unique_ptr not shared_ptr. This
will hopefully save quite a few memory barriers (for atomic ref
counting) and is cleaner.

We also prevent the copying of QueryExecutionTrees which was previously
done in several places and may be expensive but is definitely
unexpected.