Add tzdata to Dockerfile apt install packages (for timezone support)
The docker container can now be run with -e TZ=Europe/Berlin so that the
correct time is reported inside.

TODO: It would be even better, if the timezone were correct right from
the start. I did not manage to figure out yet, what to do in the
Dockerfile, so that that is a case. The tips obtained via googling did
not help.