Tweak in SortPerformanceEstimator (faster + log)
1. Right now, the estimation takes several minutes. Is that really
necessary? The main culprits are the 100,000,000 tests, so I removed
them. OK?

2. The log had two lines per sort ("Sorting ..." and "done"), it's now
only one line.

TODO: It is confusing that the sort estimation is the very first thing one
sees in the server log upon startup. The server initialization comes
afterwards. Is realize that there are reasons for this in the code.
Nevertheless, is it possible to do this the other way round?