Disable TwoColumnJoin when not supported
The TwoColumnJoin implemenentation works only in certain cases (when one of the inputs has exactly two columns).

With this fix the TwoColumnJoin is only chosen by the query planner when it is actually supported. In all other cases the more general MultiColumnJoin is used.

Also the TwoColumnJoin and MultiColumnJoin have no implementation for "Dummy Joins" with three variables, but were previously chosen for dummy joins by the query planner. This lead to errors during query processing. This commit moves these errors to the query planning and adapts the corresponding unit tests. (They were useless as they expected a certain form of query plans for queries which cannot by executed by QLever currently).