Make test is a separate line in the Dockerfile. (#428)
Put "make test" in separate line in Dockerfile. This makes it easiert to comment it out and Docker can cache the result from the compilation in the previous line.

Also comment out the SortPerformanceEstimatorTest because it takes long and fails occasionally for no good reason (e.g., if a sort takes much longer because something else is running on the machine), which is annoying.