Implemented a ConstexprSmallString
It is a simple string type, templated on its maximal size, which can be used in a constexpr context
(e.g. as a template parameter).
We can probably get rid of it again once we have support for constexpr std::string (which is standardized, but not yet implemented by the compiler vendors)