Use TwoColumnJoin only when it is actually supported (#401)
The TwoColumnJoin implemenentation works only in certain cases (e.g. when one of the inputs has exactly two columns).

With this fix the TwoColumnJoin is only chosen by the query planner when it is actually supported. In all other cases the more general MultiColumnJoin is used.

Also the TwoColumnJoin and MultiColumnJoin have no implementation for "Dummy Joins" with three variables, but were previously chosen for dummy joins by the query planner. This lead to errors during query processing. This commit moves these errors to the query planning phase.

Co-authored-by: Johannes Kalmbach <johannes.kalmbach@gmail.com>