Add easy support for building SPARQL+Text indexes
There is now a variable WITH_TEXT which default to empty. When empty,
index build and engine start are without a text component. When not
empty, a text index is built (in addition to the regular index), and the
engine is started with that extended index.

There is also a target text_input_from_nt_literals, which given an NT
file produces a wordsfile and a docsfile that if taken as input for an
index build will effectively support keyword search in literals.

Also added scientists.settings.json for the scientist collection example
on git:docs/sparql_and_text.md .