Better block boundaries for the text index so that prefix search at least works for all ASCII-only prefixes (#537)
The basic idea of the blocks of the text index is that for a given prefix length (the default is 4) there is one inverted list for all words that start with the same four-letter prefix. However, with text in arbitrary Unicode and ICU's different levels of (partial and total) orderings of strings, that notion is less well-defined than it may seem at first. For example, if the text vocabulary contains viæ (with æ being equal to ae on ICU level 1) and words starting with viae, but not the four-letter word viae, than in the version of the code so far, viæ and the words starting with viae would be in two different blocks and the prefix search for viae* failed because the code currently does not support merging multiple blocks. This is fixed now, at least for prefixes starting with four ASCII letters.

In another PR, this can and should be improved by supporting merging of more than one block.