Store the vocabulary in more compact form (#544)
So far, the vocabulary was stored as a vector of strings. This was wasteful in two respects: space (std::string has a minimum size of 32 bytes) and construction time (create an std::string for each vocabulary entry).

We now store the concatenated strings together with a vector of their offsets in this concatenation.