Add support for (not built-in) functions
The previous commits supported build-in functions (which could be
extended by extending the grammar).

Now there is also support for function that can be specified via an IRI,
like geof:distance.

This required code for a few more visit... functions of the
SparqlQLeverVisitor class and making the prefix map available to the
functions that need it (so far, SparqlParser knew about prefix map, but
it passed an empty prefix map to SparqlQleverVisitor).