Update Dockerfile
I was executing all the steps on the Quickstart guide, but the docker container immediatly exits. I think it exits, because the entrypoint is bash and I was able to run the container by commenting out the bash entrypoint and uncommenting the line above. I don't know what this entrypoint does exactly but it worked for me.