Externalization rules from settings.json now also work for literals (#655)
So far, the "prefixes-external" specified in the settings.json were only
checked for IRIs. Now they are also checked for literals (except for
literals with an XSD type and for numeric literals). For example, this
makes it easy to externalize all literals that start with "POLYGON(". On
the side, this commit improves the implementation for isXsdValue (now with a
compile-time regex) and adds a test for it (there was none so far).