Better cost estimate for non-prefix FILTER REGEX (#656)
The previous cost estimate was `std::numeric_limits<size_t>::max()`, which seemed meaningless, since cost estimates of operations are added up. The effect was that an expensive REGEX FILTER was scheduled early in the query processing (for large results) rather than later (on results that were already reduced by other joins). Now the cost is 10 times the size of the result to be filtered, which is still rather ad-hoc, but much better than before.