Build text index from all literals (optionally) (#659)
Support for option --words-from-literals for building a text index from all literals in the internal vocabulary. This can be combined with the already existing -w and -d options for building a text index from explicitly given text records. Used the occasion to improve the log of the text index building (now consistent with the log for the KB index) and several minor improvements in the code and in the documentation.