Finish implementation of GET and POST support
1. There is now proper support for GET and POST queries. The POST
queries now also allow both variants specified in the SPARQL 1.1
standard: with a URL-encoded query string ("query=...") and with content
type "application/sparql-query" and just the unencoded SPARQL query in
the body of the POST request.

2. Improved the URL parsing helper functions along the way. Eventually,
this should be done with a propery library, but until we are there,
let's improve what we have.

3. Used the opportunity to add a command-line argument for specifying
the access token to be used for restricted API calls like
clear-cache-complete. This was simple enough and completes work from an
earlier PR.