Address comments from Johannes' review and more
1. If something goes wrong during query processing as far the URL
parameters is concerned, the error message in the log and the error
message returned to the client are now in sync.

2. Better logging also at other places, in particular, when processing
the commands.

3. If the server is started without --access-token, there is now no
access to the restricted API calls at all (no matter which value is
provided in the URL parameters for access-token).

4. When an access token is provided and it doesn't match or the server
was started without --access-toknen, the whole request is now ignored.