Address Johannes' comments and improve log in various places
Note: The log was quite inconsistent regarding what it logs and what it
sends to the server. Made that more consistent now. Also logging some
more relevant information to the INFO log, like the total time for
processing a query or the basic parameters of a request (method, content
type, accept hearders). I tested it for a number of queries and it
worked alright.