FINAL round of improvements, please have another look
1. When access to a URL parameter is denied or when there are URL
parameters that were not processed, now abort the query processing with
a 400 error. In particular, this has the consequence that a file is only
served if there are no URL parameters (which makes sense).

2. In processQuery, now do the query planning *after* determining the
media type. Before it could happen that the query planning took some
time and then the processing aborted because of an invalid media type,
which didnt't make sense. Log the time required for the query planning.

3. Support an explicit "ping" action with an optional message. This is
useful for the qlever script and beyond. For example, the ping from the
qlever script is now logged in a way that we can see that it came from
the qlever script.

4. Make sparql-results+json the default media type, as required by the
standard. This is possible now because the latest version of the QLever
UI explicitly asks for qlever-results+json for every SPARQL query it
sends.

5. ALso log the size of the result when it has been computed. This
useful information was missing from the INFO log so far.

6. Minor reformatting of a few strangely formatted comments that I
encountered on the way.