Lower initial size of STXXL disk from 1 TB to 1 GB
It is weird that the initial size is so huge, even when only a very
small portion of it is actually needed to build most indexes. It is not
an actual problem on systems that support sparse files, where only the
space that is actually used is taken from the the filesystem. But some
older systems do not support sparse files. Note that STXXL will
automatically increase the size of the disk when it needs more space, so
an initial size that is too small is not a problem.

Along the way, formatted some of the other constants as, for example,
1'000'000 instead of 1000 * 1000. Some of the constant used to be (small
multiples) of powers of 2 and are now (small multiples of) powers of 10.
Does that matter anywhere?