Store and display git hash used to compile QLever (#734)
Use CMake magic to let the QLever server and index builder know about the git hash of the compiled code. Show this git hash in the first line of the log when starting the server or the index builder. When building an index, store this git hash in the index metadata and display it when starting a server with that index.