Clean up the query planner code further (#754)
Get rid of a lot of code duplication. Let the query planner find out itself that a SCAN followed by a sort is not a good idea (use a scan with the right order instead). Remove the code for the two-column join because it was not more efficient than the code for the multi-column join and probably not even correct in all cases.