Implemented the ?s ?p ?o case again.
TODO:
We actually need both: COUNT(DISTINCT) and COUNT(not-distinct), the latter one is required for fixed subjects.

TODO(2):
Unit tests. This whole business only relies on the e2e tests, which do not know, whether the pattern trick was applied or not. We should e.g. explicitly check, that the query result is the same with and without the pattern trick.