Add some builtin function: STR, STRLEN, DAY, MONTH, YEAR, RAND
The implementation of the RAND function is hacky: It is implemented as a
LiteralExpression<Variable> with special variable name "?RAND", which
then gets special treatment in the getCacheKey function (the cache key
must also be random so that repeated query execution gives a different
result every time) and in Bind::computeExpressionBind .

TODO: The logic for writing a column of the result table with the values
of a variable also occurs at other places (which do not yet work for
"?RAND"). There is some code duplication going on here, which should be
removed.