Separate LocalVocab class with basic functionality
The LocalVocab used to be an alias for std::vector<std::string> (defined
inside of the ResultTable class). New words were pushed to this vector
without checking whether they were already in the vocabulary (exception:
when a sequence of identical new words were encountered in a SPARQL
expression result, the words was only added once per sequence).

The basic data is still a std::vector<std::string>, but there is now a
phase, started by LocalVocab::startConstructionPhase() and ended by
LocalVocab::endConstrutionPhase(), where new words can be added and a
words-to-ID map records which words have already been added. After that
phase, the map is deleted and the local vocabulary is read-only. Check
test/LocalVocabTest for how it works in prinicple.

This is also used by the Values class in PR #793 (and I tested it
already quite extensively and it works).