Refactor the code handling the local vocabulary (#820)
The local vocabulary used to be a simple std::vector<std::string>, to which new words (that were not part of the normal vocabulary) were pushed. There is now an own class LocalVocab, which maintains a set of words and their local indices and supports efficient lookup both by word and by ID. There is also a method for merging two local vocabularies, which so far only works if one of the two is empty (which is a frequent use case). This is used for the join operation, which so far did not propagate its local vocabulary.