Code for communicating with client via HTTP or HTTPS
Add a new class HttpClient which can open HTTP and HTTPS connections to
a given host, and which can then send GET or POST requests, receive the
result synchronously, and write it to a (potentially very large) string.
Uses Boost.Beast, just like our HttpServer.

Add a combined test for our HttpServer and HttpClient (for a meaningful
test one needs a server and a client). So far only tests the HTTP case
because our HttpServer does not do HTTPS yet.

Used the occasion to rename util/HttpServer to util/http because
httpServer was simply a misnomer, which became even more apparent with
the addition of HttpClient. This change required many small changes (in
includes and in the various CMakeLists.txt).