Code for communicating with another host via HTTP or HTTPS (#823)
Add a new class HttpClient which can open HTTP and HTTPS connections to a given host, and which can then send GET or POST requests, receive the result synchronously, and write it to a (potentially very large) string. Uses Boost.Beast, just like for our HttpServer. Includes code to parse a URL into its components (protocol, host, port, target).
    
Add a combined test for our HttpServer and HttpClient (for a meaningful test one needs a server and a client). So far only tests the HTTP case because our HttpServer does not do HTTPS yet.
    
Use the occasion to rename the directory util/HttpServer to util/http because httpServer was a misnomer already before, which became even more apparent with the addition of HttpClient. This change required many small changes (in the includes and in the various CMakeLists.txt).