Efficient COUNT DISTINCT for full index scan
The following kind of query is now also processed more efficiently, avoiding an actual scan of the whole index. This continues work on the efficient processing of typical statistics queries from #824 and #830.

SELECT (COUNT(DISTINCT ?s) AS ?count) WHERE { ?s ?p ?o }

The argument of the COUNT DISTINCT can be either ?s or ?p or ?o. The respective counts are precomputed and stored as part of the index meta data. They are shown under "Index Information" in the QLever UI. The internal triples are NOT considered anymore in these counts (but there are separate counts for those in the index meta data, too).

IMPORTANT: This requires rebuilding the index (or manually adjusting the .meta-data.json file from the index build). It also requires a code change in the QLever UI because the variable names in the JSON file have changed.