Made a pass over the comments
1. A bit of rewriting and reformatting, nothing major. No change of
code.

2. Shortened the long comment in ColumnBasedRow.h because much of it is
now already explained in the comment in front of IdTable.h. Referred to
that comment now.

3. TODO: The subdirectory `idTable` contains only a single header file.
I would move that header file up again and remove the subdirectory.
Also, as discussed, rename the two files to `IdTable.h` (merge with the
existing file with that name) and `IdTableRow.h`.

Thank you for this impressive PR.