Change the `IdTables` growthFactor to 1.5 and also amortize the costs of repeated `resize` calls.
A growth factor of 1.5 was also present in the old row-based `IdTable` and is
considered to be better than a factor of 2 also by many industry experts (see for example `folly::fbvector`.

A `resize` operation now increases the capacity by at least the growth factor.
This also amortizes the costs of a chain of "small" `resize` operations.