Propagate the local vocabulary in all operations (#838)
The local vocabulary is now properly propagated in all operations. In the GroupBy operation, a copy is made before adding words to the local vocabulary (otherwise the local vocabulary of other operations would be affected, and there would be unexpected results when reusing results from the cache). There are tests that the propogation indeed works as expected for each operation.