Refactor the `Timer` class (improve interface and use std::chrono) (#871)
This was a very old class copied from the CompleteSearch code. The new interface is more meaningful for QLever. In particular, it is no longer necessary to stop the timer to get the current duration (which now has its own type Timer::Duration and can be converted to the desired unit of time). Also, the implementation now uses std::chrono instead of the old gettimeofday. There are now also proper unit tests.