Add a Github action that uploads a given code coverage file. (#877)
This workflow file `upload_coverage.yml` will always be taken from the master branch (not from the pull request) and thus always has access to the stored secrets.

This is the first step towards avoiding the
spurious failure for the Codecov uploads.
The second step (measuring the code coverage
for a PR and sending it to the `upload-coverage` workflow) will be done
in a separate PR.