Improve Values and ResultTable classes and related code
Clean up and improve the `Values` class. Most notably: in `writeValues`,
move the values out of `parsedValues_` to avoid unncessary copies. This
necessitates a change in `test/LocalVocabTest`. Write a unit test for
the `sanitizeValues` function.

In the `ResultTable` class, make `localVocab_` private and provide a
public getter and setter and use them everywhere. Also improve the
documentation of the class.

Add an algorithm for removing the duplicates in a given `std::vector`
without changing the order of the elements.

Several associated improvements in the code and also some minor
unrelated stuff I stumbled upon.