Column-based storage of the indices on disk (#862)
The representation of the index files on disk is now also column-based (after #841 already changed the layout of the central `IdTable` class from row-based to column-based). The index files are a bit smaller now (13% for Wikidata) because single columns, stored contiguously, can be compressed better than sequences of pairs (as in the row-based layout so far). Index building time does not seem to be affected. Extensive unit tests with perfect coverage of the code in the inner functions.