Each permutation of each index now has its own block cache (#889)
The block cache is used to speed up our end-to-end tests, which often read the same blocks and then only use a small portion of such a block for a single query. So far there was one global (singleton) block cache. This had the potential to cause nasty bugs when different tests created and used multiple indexes. This is now done properly, with one block cache for each permutation in each index.