First working implementation of SERVICE
New version rebased to the current master and with additional tests.

The improvements of the handling of `Values` and `LocalVocab` (PR #880)
and the improvements in `http/util` (PR #900) have already been factored
out. So what is remaining here directly pertains to the implementation
of SERVICE.

TODO: Tests in `test/SparqlAntlrParserTest.cpp` are still missing, maybe
Johannes can help me with that.