Improve handling of `Values` and `LocalVocab` (#880)
Improve code for `Values`: In `Values::writeValues` move the values out of `parsedValues_` to avoid unncessary copies; allow UNDEF values in the parser and remove the then no longer needed `sanitizeValues` function; make sure that most of the (so far largely untested) code is covered by unit tests.

Improve how a `LocalVocab` is shared among results: `ResultTable::localVocab_` is now private and there are the following new methods for sharing, copying, or accessing it: `shareLocalVocabFrom`, `shareLocalVocabFromNonEmptyOf`, `getCopyOfLocalVocabFrom`, `LocalVocab`, `LocalVocabNonConst`. There is now a safeguard against adding to `localVocab_` once it has been shared.

Add an algorithm for removing the duplicates in a given `std::vector` without changing the order of the elements.

Several associated improvements in the code and documentation and some minor unrelated stuff I stumbled upon.