Own type `TripleComponent::Literal` instead of `std::string` (#913)
So far, our class `TripleComponent` stored a parsed literal as a `std::string`. The class already had own types for `Variable`, `UNDEF`, and double or integer values. Now it also has an own type `Literal` (which is is implemented using a new class `NormalizedRDFString`). It remains to also add own types for IRIs and blank nodes.