New version with a first proper test
There is now a test that checks for all existing triples whether the
found location is correct (by checking `id2` and `id3` at the found
position in the found block, note that the block does not have an
explicity `id1` for a given position).

The `findTripleInPermutation` method is still (very) inefficient in that
it goes through the complete relation metadata in order to find the
sequence of `id1`s relevant for a block. This will be fixed in the next
commit.

Note: the previous commit lacked the new files `DeltaTriples.h`,
`DeltaTriples.cpp`, and `DeltaTriplesTest.cpp`.