Fix block size bug (#915)
The block sizes for the OPS, OSP, SOP, and SPO permutations were too small because of a bug in how it was determined when to end an block. For example, for the current Wikidata index as of this writing, the PSO and POS permutation have 48,769 blocks each, but the SPO and SOP permutations have only 1967 blocks each. This bug is fixed now.

Co-authored-by: Johannes Kalmbach <johannes.kalmbach@gmail.com>