Compute `UndefStatus` of each column during query planning (#918)
During query planning, compute an `UndefStatus` for each column of each (intermediate) query execution tree. The status can be either `AlwaysDefined` (we are certain that the column does not contain any `UNDEF` values) or `PossiblyUndefined` (we cannot exclude the possibility that the column contains `UNDEF` values). This information is not yet used anywhere, but we plan to use it to implement a correct optional join as efficiently as possible. In particular, an optional join can be expensive to compute when multiple join columns contain `UNDEF` values and is much cheaper to compute when only one of the join columns involved contains `UNDEF` values.

Also add a method to `ResultTable` that computes the number of occurrences of each datatype in each column during query time. This is currently computed brute force (by iterating over each column) and only used with `-DENABLE_EXPENSIVE_CHECKS` enabled (see #928) to check at query time whether the actual number of `UNDEF` values in a column is consistent with the aforementioned `UndefStatus` computed during query planning.