Correct handling of LIMIT and OFFSET for subqueries (#931)
For subqueries, `LIMIT` and `OFFSET` were incorrectly ignored so far. This is now fixed, and in particular fixes #929. QLever still computes the full result before applying the limit or offset value (except for the "full index scan", which has a special implementation when combined with a `LIMIT` clause), but only the truncated result is stored in the cache. That is, when asking the same query again with a different `LIMIT` and `OFFSET` (or none at all), the result has to be recomputed.

`LIMIT` and `OFFSET` at the very end of a query are handled just like before, and different than for subqueries. Namely, the full result is computed and cached (again, except for the above-mentioned special case), and the `LIMIT` and `OFFSET` are considered only when exporting the result.

As an add-on, the runtime information now has an own entry for `LIMIT` and/or `OFFSET` (one entry for both) if they are specified.