Method `mergeTriples` with good unit tests
There is now a method `LocatedTriplesPerBlock::mergeTriples` that merges
the delta triple into a given (possibly partial) block. There are unit
tests for the following cases: full block with unrestricted `id1` and
`id2`, full block with restricted `id1`, patial block with restricetd
`id1`, partial block with restricted `id1` and `id2`, and the latter
with only a single column.

Removed the previous complicated version that tried to do it in place.