Delta triples merged for both scan types now
The variant of `CompressedRelation::scan` with two `Id`s fixed now also
considers delta triples. This was significantly more complicated than
for the variant with only one `Id` fixed and required quite a bit of
refactoring.

TODO: For the first incomplete block when only a single `Id` is fixed,
delta triples are still not considered. That should be easy to add
though.