Delta triples now considered for all blocks
There was still one case missing: the possibly incomplete bloc at the
beginning when only a single `Id` is fixed. Now delta tiples ae also
considered for these blocks.

TOOD: The unit test works for a permutation with a single relation, but
there is still a problem when there are multiple relations.