Fix the CSV export of SELECT queries.
The first line is now
var1,var2
instead of
?var1,?var2

Entities are serialized as
entity
instead of
<entity>

And strings as
someLiteral
instead of
"someLiteral"@en-ca

The only exception are Literals or entities that have to be escaped, they will be exported as
"literal, with comma".

This is now according to the SPARQL and CSV standard.