Correct printing of benchmark table with zero rows (#964)
In certain cases it can happen, that a benchmark table result is empty, consider for example that a benchmark allows for the configuration of a memory limit, and none of the measurements can be performed with the specified
limit. Previously this case crashed. Now empty tables are correctly printed with an explicit note saying that a table without any rows was encountered.