Correct handling of IRIs and strings in expressions and CSV export (#962)
The CSV format now correctly implements the SPARQL standard according to
https://www.w3.org/TR/sparql11-results-csv-tsv/. Most notably, literals are serialized without
language tags and datatypes, and typically without quotation marks (see below). IRIs are serialized without angle brackets. Quotation marks are only used to escape IRIs or literals according to RFC4180, but not to distinguish IRIs from literals. This fixes #911.

Also SPARQL expressions that work on strings now only see the inner content of a literal or IRI (no datatypes, language tags, quotation marks, or angle brackets). For example, the result of a
GROUP_CONCAT that previously was "lit1"@en, "lit2"@fr is now correctly lit1, lit2.

For REGEX expressions, one now has to explicitly use the STR function when also wanting to search IRIs and numeric values. As long as the call to STR is directly inside the regex expression, this is implemented efficiently. See the following two examples:
REGEX(?x, "bla") matches "blabla"@en, but not
REGEX(STR(?x), "bla") matches both of the above.
BIND (STR(?x) as ?strx)
REGEX(?strx, "bla") matches both, but is inefficient, because the resulting strings of the STR
function are always fully materialized.

The following related issues are still left open:

A similar handling of the STR function should be used in all other functions that work on strings. For example, currently the STRLEN function currently always behaves like STRLEN(STR(...)).
Prefix regexes are currently unable to match numbers and other values that are encoded in the ID, and also are unable to match entries in the local vocabulary.