`Permutation` class no longer templated (#965)
So far, the `PermutationImpl` class was templated on the comparison that sorts triples according to the permutation, and the underlying type of `IndexMetaData` (in memory for PSO and POS, on disk for SPO and SOP and OSP and OPS). The first template parameter has been unused for quite some time and is now removed. The second template parameter is now also removed, with the consequence that the metadata for PSO and POS are now also stored on disk.

The new `Permutation` class is now much easier to use, since we do not have to template every function that uses it anymore.

IMPORTANT: This code change requires a rebuilt of the index. We should monitor if the new performance overhead for POS and PSO turns out to have a significant performance overhead. We don't expect this, but if it does, we can add a cache for `IndexMetaData`.