Make QLever compile with Clang 16 and libc++ using Conan (#969)
QLever can now also be compiled with Clang 16 *and linked to LLVMs libc++*. This requires an installation of Boost that is also linked against `libc++`. On Linux systems this is best installed via the Conan package manger. This toolchain (Clang 16 + libc++, Boost via conan) is used for the Codecov workflow, so you can look at the `code-coverage.yml` for an example of this workflow.
Some small changes were required with QLever to make it compile with Libc++:
* The `restricted` proxy reference of the `IdTable` class cannot be used because of strict requirements of libc++ for `std::iterator_traits`.
* `absl::from_chars` has to be used instead of `std::from_chars`
* `std::integer_sequence` cannot be used for non-integral types like `std::array<..>`, so we use a drop-in replacement `ad_utility::ValueSequence`
* Some other minor changes.

In the same go we have also dropped the support for Clang 13, 14, and 15. This means that all our supported Toolchains now support the lazy views from the `<ranges>` header.