Use own type `WordVocabIndex` for words from the text index (#973)
So far, the type `VocabIndex` was used for entities and literals from the RDF input data, as well as for words from the text index. These are now two separate types: `VocabIndex` and `WordVocabIndex`. On the side, cleaned up the code and removed some unused functions. All this is preparation for improving the text search code, which is rather old and inefficient in various places. 

TODO: We should eventually rename `VocabIndex` to `RdfVocabIndex`, but we leave this to a separate PR because this affects many files and there are currently many PRs open and we don't want to annoy the people working on them.