Remove inverted lists for entities from text index  (#979)
In the code for the text index so far, there was, for each entity, an inverted list that contained the text records where that entity occurred. Because of this, there was a `TextBlockMetaData` object for each entity, of size 96 bytes each. These objects were kept in RAM (both at indexing time and at query time), which for Wikidata with text search in literals were around 40 GB. In an old version of the code, these inverted lists for entities allowed efficient queries that find all text records containing a given fixed entity, but that was dysfunctional for some time already.

These inverted lists and the associated `TextBlockMetaData` objects and the associated RAM overhead are now gone.