Consistently use the `ColumnIndex` typedef (#990)
QLever currently contains a type `ColumnIndex` that is just a typedef for `uint64_t`. So far, sometimes `size_t` was used where `ColumnIndex` would have been semantically more appropriate. This helps twofold: It improves the readability of the code, and it brings us closer to a successful compilation on MacOS, where `uint64_t` and `size_t` are different types. Note that in the future we might use a stronger type for `ColumnIndex`, but that would require changes in many additional places.