Turn WARN into DEBUG messages when date information is discarded
It turns out that there are very many of these also for relative clean
datasets like Wikidata. We don't want the normal index build log to be
full of warning messages.

Eventually, there should be an option to have such messages written to a
separate file. But that is work for a separate PR.