Don't terminate the server when socket operation fails (#996)
When a socket operation was aborted, the closing of the socket threw an exception due to a subtle bug. That exception was then not caught and the server was terminated. The bug is now fixed and exceptions that would lead to a termination of the program are now suppressed by handling the cleanup with our own `ad_utility::makeOnDestructionDontThrowDuringStackUnwinding` instead of with `absl::Cleanup`.