Make QLever run on macOS (#999)
Since this commit it is possible to compile and run QLever on `macOS` using LLVM-Clang 16 with dependencies installed via Conan. There also is a GitHub actions workflow which continuously checks this toolchain and also serves as a documentation on how to make QLever run on mac. Aside from adding the workflow this commit only contains two small changes in the actual code base:
* A small bugfix in the `MmapVector` class for code that is only used on non-linux systems.
* All the tests for the `Timer` class are deactivated on macOS, as `std::this_thread::sleep_for()` doesn't seems to be 
  precise enough on these systems.