Store first and last triple in block metadata + a block is now an `IdTable` (#1004)
The metadata for a block now stores the first and last triple completely (so far, only a subset of the `Id`s were stored). When using this new code, indexes must be rebuilt.

On the side, improve and clean up the code for reading blocks from a relation. In particular, a block is now a regular `IdTable` instead of the internal `columnBasedIdTable::IdTable<Id, 0>`. This requires that classes like `Permutation` and `CompressedRelationReader` now have an allocator. The advantage is that these blocks are now considered when checking whether the total amount of memory used by QLever for query processing is exceeded.