Benchmark groups can now contain tables (#1006)
The Benchmark modules allows us to group benchmarks in groups (a set of measurements that are associated with a common name) and tables (a more structured set of measurements that is grouped in a certain layout of rows and columns). Now such tables can also be members of groups which makes organizing large sets of benchmarks more convenient.