Also parse empty configurations in BenchmarkMain (#1026)
The configuration for a benchmark might have mandatory options without a default value that therefore always have to be specified. To correctly handle this case, it is crucial that the configuration is parsed even if it is empty because no configuration file or shorthand was specified. This was previously not done, but as of this commit, the configuration is always parsed.