Native build with thread sanitizer + fix deadlock in `ThreadSafeQueue` (#1027)
Add another native build that uses the thread sanitizer (instead of the address sanitizer, the two are mutually exclusive). For the MAC build, use neither because it takes too long otherwise.

While at it, fixed two thread-safety issues. First, there was a potential deadlock in the `(Orderd)ThreadSafeQueue` when accessing the `finish_` variable, which used to be `std::atomic` and now isn't any longer (that was a significant bug). Second, the access to `timer_` in `BatchedPipeline.h` was not thread-safe (this was pointed out by the thread sanitizer, but never made problems so far).

On the side, improve the logging in the function `terminateIfThrows` (which is only called in absolute Wurst-case scenarios).