Deterministic sort order of join columns in multi-column joins (#1029)
For multi-column joins (which can occur for the QLever operations `MultiColumnJoin`, `OptionalJoin`, and `Minus`), make sure that the sort order of the join columns is deterministic. In particular, this is important when retrieving cached results for such queries. 

In the previous version of the code, one of our AC queries was a multi-column optional join, the result of which was pinned to the cache. However, reusing that result often failed because the sort order of the join columns computed by the query planner was not deterministic.

On the side, cleaned up and simplified the code surrounding this change. As a side effect, QLever now allows joins of the full index with itself, like `?s ?p ?o1 . ?s ?p ?o2`.