Version for QLever's index. (#1035)
QLever now has versioning for its index format. That way, users now get a proper error message when the format of the index is not compatible with the version of QLever used to start the server. So far, a cryptic error message appeared, typically an assert failure.

IMPORTANT: When a future commit changes the index format, the corresponding entries in IndexFormatVersion.h have to be updated.

* Try if this check now works as it should.

* Try to debug this workflow.

* Trying to fix the index versions.

* Fix two code smells.

* Changes from a review with Hannah.

* The current index version is now there.

* Renamed almost everything.