Versioning for QLever's index format (#1035)
QLever now has versioning for its index format. That way, users now get a proper error message when the format of the index is not compatible with the version of QLever used to start the server. So far, a cryptic error message appeared (typically an assert failure), which left the user puzzled as to what they did wrong.

IMPORTANT: When a future commit changes the index format, the corresponding entries in `IndexFormatVersion.h` have to be updated. To help remember this, there is a new workflow `check_index_version.yml`, which checks that if these entries have not been changed, the end-to-end test runs through without errors when building the index using the current master, yet starting the server using the code from the designated commit.