Add a Strong Type for Memory Sizes (#1032)
Add a class `ad_utility::MemorySize` that stores an amount of memory in bytes in a strongly typed manner. It can be parsed from strings (e.g. "4 kB" or "8.3 GB") and set from user defined literals (`4.34_GB` etc.). This commit does not yet use the `MemorySize`, but this is left for following PRs.