Implement the SUBSTR function (#1057)
The first implementation of a non-trivial string function. `SUBSTR` is a bit more challenging than the others because it takes a string argument as well as numeric arguments and because the numeric arguments have default values and there are many non-trivial corner cases to consider for implementing the standard correctly.