Allow an empty index (#1061)
So far, the index builder crashed, when no triples were found in the input file. Now the index is built with a warning, but otherwise behaves as expected. A SPARQL endpoint with an empty dataset is used frequently in the SPARQL 1.1 Test Suite, where the data is provided in the SPARQL queries using VALUES clauses.