Show the completed words when using `ql:contains-word` (#954)
When the fulltext index is used (via the special `ql:contains-word` predicate) and the specified word is actually a prefix (for example `auto*`), then the completed words for each match (e.g. `automatic` or `automotive` are now also shown in the output.
In detail:
For each prefix that is part of the object of `ql:contains_word`, and additional column will be added to the result. For example the triple `?q ql:contains-word "auto* albert rela*"` will add one column for the matches of `auto*` and another column for `rela*`. If a text in the result matches the same prefix for multiple words, then the cartesian product of all the matches will be materialized in the result. For example (given the triple above) if a text contains the words "albert automobile automatic relative and relational", this text will lead to four result rows (2 * 2 because we have two matches for `auto*` and two for `rela*`.
The names of the added variables are derived from the subject variable of the `contains-word` triple as well as the typed prefix. They can be exported via `SELECT *`. Their exact format might still change in the future.