Simplify test port selection (#1102)
The test code that needs to start a Server on an available port (and we don't care about the exact port number) now asks the operating system for such a port. This is done by specifying the special port 0 when constructing the server. After the construction we can then query the server which port is actually used. This simple code replaces a manually implemented random sampling of ports.