Deterministic sort order for literals + do not externalize ql: IRIs (#1098)
The sorting of the vocabulary was nondeterministic for literals that only consist of punctuations characters like `"..."@en`. This  affected the merging of the partial vocabularies, which demands a consistent total order on the items. As a consequence, for datasets with many literals involving special characters in many scripts (like Wikidata), it could happen that the same literal was assigned multiple IDs. The issue is now almost resolved except for a few very obscure and rare cases. The index builder now reports all pairs of consecutive vocabulary items that violate the total sort order. For Wikidata, three such pairs are reported. Only queries involving these items would be affected if at all.

On the side, now ensure that QLever-internal IRIs (like `ql:langtag` or `ql:has-predicate`) are never externalized. Previously, when they were externalized due to the `settings.json`, they did not work when they were explicitly present in a query.