Compressed external `IdTable` storage and sorting (#1090)
So far, the index builder used a `stxxl::vector` to store the triples and then `stxxl::sorter` to sort the various permutations. This used a lot of space: the vectors are uncompressed (hence using 24n bytes for storing n triples), there is the vector-typical overallocation (of around 50%), and there is another factor of two because the sorting is not in place. For example, when building the index for Wikidata, this resulted in a space consumption of around `2 TB` only for the `stxxl` files. This comes on top of the actual index files produced, which have a total size of slightly less than `1 TB` in the end.

The input triples are now stored and sorted *compressed* using two new classes `CompressedExternalIdTable` (for storing an `IdTable` compressed on disk) and `CompressedExternalIdTableSorter` (for sorting an `IdTable` compressed on disk). This reduces the space consumption of the files by roughly a factor of 10, so that the space for storing and sorting the triples is now only a fraction of the total size of the index files produced. In particular, Wikidata can now be build with around `1 TB` of disk space. The time for building the index is essentially the same as before on SSDs, and significantly faster than before on HDDs.

NOTE: While testing this PR, we noticed that the index builder tends to use more RAM then what is specified by `--stxxl-memory-gb` (which by the way is now a misnomer). During the initial parsing phase, the peak depends on the total size of the literals in a batch, which has nothing to do with this PR. During the sorting of the permutations, the peak RSS can be quite much larger than the values specified by `--stxxl-memory-gb`, but it is currently unclear which part of that is due to `mmap` allocations that do not actually require RAM. This needs further investigation in the future.