Refactor the PropertyPath class
Prior to this change, the class used an enum with confusing members
like TRANSITIVE_MIN. It also had a limit variable, which does not make
sense with the current SPARQL standard.

This change introduces new enum members which correspond to the Property
Path Expression Names of the SPARQL standard. Simplified the handling of
modifiers (+,*,?), such that limit is no longer needed.

Type: refactoring