Refactor the TransitivePath class
Prior to this change, there was a lot of code duplication in the
TransitivePath class. This buried important logic which chooses the
starting nodes of the transitive hull algorithm. The choice of the
starting nodes is important for the performance of this algorithm.

This change introduces the struct TransitivePathSide, which holds all
relevant information on one path side. It also splits the public methods
computeTransitivePath, computeTransitivePathLeftBound and
computeTransitivePathRightBound into multiple private methods.
Additionally, some new unit tests were introduced and the old tests
refactored as well.

Note: This is a "safe" refactoring. It is supposed to keep the
functionality of the engine the same as before.

Type: refactoring