Fix a bug in TransitivePath::computeResult
Prior to this change, the computeTransitivePath function was called with
the left as start side and the right as target if the right hand side is
and Id. But in this case, the computation is faster with the right as
start side and the left as target.

This change adds a check in computeResutl to call computeTransitivePath
the correct way if right is an Id.

Type: bugfix