Adjust semantics of isBound
Prior to this change, the TransitivePath could be bound on both sides by
the QueryPlanner, which is not properly implented.
Also, the function TransitivePathSide::isBound meant that a side is
either and Id or it was bound via bindLeftOrRightSide function. The
QueryPlanner needs this function, in order to not bind the
TransitivePath object twice. But in the case of a variable and an id,
the TransitivePath::isBound function would now return true, although the
variable side of the TransitivePath could be bound.

This change adjusts the meaning of TransitivePathSide::isBound, s.t.
it is true iff the TransitivePath has been bound via bindLeftOrRightSide
function. Also TransitivePath::isBound returns true, iff either side is
bound.

Type: bugfix