Add tests to QueryPlannerTest for TransitivePath
Prior to this change, the construction of a TransitivePath from a
ParsedQuery by the QueryPlanner was untested. This includes the binding
of a TransitivePath.

This change introduces new tests to the QueryPlannerTest for the unbound
and bound TransitivePath cases. The TransitivePath class had to be
extended by getters for the left and right side, in order to test that
they were correctly constructed. The case where a TransitivePathSide has
an Id is untested as of yet.

Type: feature