Fix a bug in TransitivePath::transitiveHull
Prior to this change, there were still references to `_lhs`
and `_rhs` left in the transitiveHull function. This causes wrong
results in the case where the left side is unbound and the right side is
an Id.

This change replaces references to `_lhs` and `_rhs` by an optional Id
that can be supplied to the function. This optional can be used, if the
targetSide value is an Id.

Type: bugfix