Speed up first phase of index build (#1124)
The first phase of the index build (where the input triples are parser and the partial vocabularies are built) is now more than twice as fast. The main change is that the space for the hash maps for the partial vocabulary (one per batch) is now reused (via a new special-purpose class CachingMemoyResource.). There are also various other small improvements. For example, various unnecessary string copies from the previous code are now removed. At one place, multiple hash maps are converted to a large vector, which is now parallelized.

IMPORTANT: To make the best use of these improvements, when feeding input files compressed with BZ2, use lbzcat -n 4 instead of bzcat in order to decompress in parallel (the argument of the -n is the number of decompressor threads).