Support more empty Transitive Paths (#1122)
QLever now supports patterns like `?x <is_a> ?y. ?y <subtype>* ?z` where the transitive path `<subtype>*` can be empty, but is constrained by the join on `?y`.  For cases like this where the connection between the empty path and its neighbors in the tree is made via explicit variables (as oppsed to the short hand `?x <is_a>/<subtype>* ?y` the query planner now can optimize the join order for such queries. Consider for example :
```
?x <profession> <astronaut> . # very few instances
?x <type>     ?type .  # very many instances without the previous restriction.
?type <subtype>* ?subtype
```
The query planner now first joins the first two triples (which yields very few bindings for `?type` and then uses this small subresult to compute the transitive path.

Note that this optimization does currently not work for the follwing pattern:
```
?x <profession> <astronaut>
?x <type>/<subtype>* ?subtype
```
Here QLever first materializes the transitive path for all types of all entities. This restriction will be lifted in an upcoming commit.

This commit also completely refactor the `TransitivePath` class to get rid of a lot of duplicate code.