More correct appending of benchmark results to previously existing files (#1155)
When appending the JSON result of a benchmark run to a preexisting file that already contains the result of previous benchmark runs, the contents of the file now is a single valid JSON object. Previously it was the concatenation of multiple independent JSON object which made the results hareder to parse when analyzing them.