Run parsing and planning of queries on the threads dedicated for query processing (#1156)
QLever manages two separate thread pools: One is used for the HTTP/Websocket server and the other for computing query results. The parsing and planning of queries is now performed on the query processing thread pool. This allows to still access the Server for debugging purposes if the engine is hanging because of long running processes that exhaust the resources for query processing.

Also remove the ancient `.html/.js/.css` files that were from a rudimentary UI and also remove the ability to server arbitrary files from the Server as QLever shouldn't be used to serve arbitrary files.

Also refactor the `Server::processQuery` function by factoring several long code blocks into separate named functions.