Allow the loading of indices with two permutations (#1159)
So far, when the index was built with only two permutations, the server start failed, because the keys `num-subjects-normal`  and `num-objects-normal` were not found in the `.meta-data.json` file (because this information is obtained from computing the SPO/SOP and OPS/OSP permutations pairs). Now these two values are simply set to zero when the keys are not found in the ".meta-data.json". Fixes #1142