Add the ability to store additional columns in the permutations (#1121)
So far, QLever's precomputed permutations (PSO, POS, SPO, SOP, OSP, OPS) had two columns each. This change adds the possibility to have additional columns. On the side, the respective code (in class CompressedRelation and the classes that use it) is now nicer and simpler (nicer interfaces and removal of code duplication).

The additional columns are not yet used anywhere, except in a few unit tests. There should be no performance difference compared to the previous code.