Fix spurious ASAN errors in HttpTest (#1164)
These test failures happened, when tasks were scheduled on an io_context and these tasks referenced objects the lifetime of which was shorter than that of the io_context. The solution consists of two parts: Pass these references as shared_ptr s.t. the lifetime is always sufficient, and make sure to explicily stop the io_context at the end of test runs to make sure that the tests always can run to completion (another spurious issue).

Note that these issues only arise in unit tests, because in production, all the involved objects live through the whole lifetime of a QLever server instance.