Specify memory sizes for command-line options with units, e.g. `30G` (#1067)
Command-line parameters to QLever that specify an amount of memory now have to be passed as a `MemorySize`. Previously these were passed as plain integers and the name of the option reflected the unit. So for example we now have to write `--memory-for-queries 30G` instead of `--memory-for-queries-in-gb 30`. Additonally the `MemorySize` class is now used consistently inside QLever, for example in the `Cache` class, the `AllocatorWithLimit` and during the index building where memory limits for the external sorting are managed.