Fix bug in lazy join between blocks (#1195)
There was a bug in the lazy join when the blocks of the left operand overlapped with the blocks from the right operand. This is now fixed. Here is a query that resulted in an assert failure before and that now works as it should: https://qlever.cs.uni-freiburg.de/wikidata/t4rYnv .