Make the `QueryHub` class simpler, and try to get rid of test stalls on MacOS. (#1205)
The old logic that was used to remove a Websocket distributor from the global registry after either the corresponding query was finished, or all listening connections were closed, was rather complicated. This was because of the combination of synchronously blocking code (destructors) and asynchronous code (the server module in general).
As of this commit, all these deletion operations are performed asynchronously, s.t. destructors never have to block. The correct behavior of this deferred deletions is ensured by Boost::ASIO's scheduling guarantees and is checked by several assertions.
This hopefully gets rid of some spurious test failures on MacOS where a test would be caught in a deadlock.