Add subject patterns for O as fifth column to OSP&OPS and PSO&POS (#1177)
This continues #1168.

That PR changed the order, in which the permutation pairs are computed (new order: SPO&SOP, OSP&OPS, PSO&POS) and added a fourth column to the OSP&OPS and PSO&POS permutations for the subject pattern pertaining to the *S* in the respective permutation.

This PR adds a fifth column to OSP&OPS and PSO&POS permutations for the subject pattern pertaining to the *O* in the respective permutation. This increases the total indexing time and the total file size of the `index.???` files only by about 3% (measured for the Wikidata build).

As part of this PR, refactored the central join algorithm, which now also enables a lazy *optional* join (when there are no UNDEF values in the join columns). This is needed for adding the mentioned fifth column because not all entities have a subject pattern.