Remove functionality from File.h to fix undefined behaviour (#1225)
The function  `ad_utiliyt::File::readFile` was was subject to undefined behaviour if the buffer passed to the function does not contain a null terminator. This function is now completely removed and its only usage is replaced by `std::ifstream`.
Also some parts of the class `ad_utility::File` are cleaned up by removing unused functionality and unused includes.