Make the ISBLANK function more efficient
At the end of Vocabulary::readFromFile, compute the ranges of IRIs,
blank nodes, and literals in the internal and external vocabulary and
remember them.

Implement an `IsBlankNodeValueGetter` that simply checks if the `Id` is
of type `VocabIndex` and if yes check whether it is in one of the
computed ranges. This is much faster than the previous implementation,
which looked up the string for each `Id`.

TODOs:

1. Remember the `Id` ranges instead of the `VocabIndex` ranges. That
   way, we don't have to extract the `VocabIndex` from the `Id` for each
   check (which is pretty fast though, just masking out some bits).

2. Also implement this for ISIRI and ISLITERAL. Avoid code duplication
   by having only one `...ValueGetter` for all three. Or at least a
   common implementation.

3. Clean up the code, it's still a but ugly. But it works!