Write vocabulary files to separate directory
The vocabulary files can now be written to (and read from) a separate
directory than the other index files. This directory can be specified in
both `IndexBuilderMain` and `ServerMain` via the new command-line option
`--vocabulary-basename` or `-v`. The directory for the index files is
specified like before via `--index-basename` or `-i`. If no directory
for the vocabulary files is specified, the same directory as for the
index files is taken (that was the status quo before this PR).

This is useful for datasets with a huge vocabulary. For example, the
vocabulary files for UniProt have a total size of over 3 TB (mostly due
to `.vocabulary.external` and `.vocabulary.external.idsAndOffsets`).