Automatically annotate each benchmark result with its time of creation. (#1221)
The  metadata of each benchmark run now has a default `time-of-measurement` entry, that is automatically set to the date and time of the benchmark run. 
This makes it easier to distinguish different benchmark runs from each other.
Also make the serialization of the benchmark results use `nlohmann::ordered_json` instead of `nlohmann::json`, such that all the members will be consistently serialized in the order that they were added.