Fix prefix search for words in external vocabulary
So far, the method `Vocabulary::prefix_range` returned a range of
indexes of words in the internal vocabulary that match a given prefix.
This is now replaced by a method `Vocabulary::prefixRange`, which
returns two ranges, one for the internal and one for the external
vocabulary. This can easily be extended to more ranges when needed.

Based on this, our efficient implementation for REGEX when the regular
expression is a prefix now finds items in both the internal and the
external vocabulary (so far: only in the internal vocabulary).

On the side, identified and fixed a bug in the previous code, where the
special predicates starting with `@` would not be found if
`prefixes-external` matched `@`. As a conesequence, QLever now works as
it should even with `"prefixes-external": [""]` (that is, everything
expect QLever-internal predicates, ends up in the external vocabulary)
in the `.settings.json` file. In conjunction with #1223, this makes it
now possible to start a QLever server with very little RAM.