Efficient GROUP BY for counting number of objects or subjects
Queries of the following kind currently take very long or even fail on
QLever when the predicate has many triples:

```
SELECT ?object (COUNT(?object) AS ?count) WHERE {
  ?subject rdf:type ?object
} GROUP BY ?object
```

The reason is that all triples from that predicate are first loaded into
RAM and are then iterated over.

This PR processes queries of the form above by just looking at the meta
data of the respective permutation and only loading and decompressing
blocks when a block contains triples with more than one `col1Id`.

TODO 1: That last part is currently still missing, such that the current
result is approximate. However, it's good enough for object AC queries
without context already. The special processing is currently only done
when a predicate spans more than 100 blocks.

TODO 2: Just a proof of concept for now. The code is still ugly and
there are no tests for the new code yet.