For small index scans, avoid spurious copy of whole block (#1260)
For small index scans, which read only a part of a single block, QLever so far made a copy of the whole block, even when the block was already in the block cache. This spurious copy is now avoided, so that, when the block is already in the cache, only the requested part is copied. In particular, this makes full index scans much faster and also helps other queries with ?s ?p ?o where ?s is bound by the rest of the query.

Note: There is potential for more performance gains for such scans, but we have other issues for now.