Fold blank node identifiers into the `Id`s (#1294)
Blank nodes used to have an entry in the vocabulary, like normal literals. Now they are `Id`s with an own datatype and their identifier (a unique integer) as part of the `Id` (just like we already do it for integers, floats, and dates). This makes the code cleaner, reduces the size of the vocabulary (significantly for some datasets, like dblp), and speeds up the function `ISBLANK` (which only has to check the datatype bits now).

In an unrelated change, improve the test coverage for the text index.