Customizable compression of both internal and external vocabulary (#1296)
So far, the internal vocabulary was prefix-compressed (with single-byte codes), while the external vocabulary was always stored uncompressed. Now both internal and external vocabulary can be compressed. The compression/decompression scheme is easily customizable: it's simply a template parameter of the new `CompressedVocabulary` class. Three schemes are currently implemented: the above-mentioned prefix compression, FSST, and FSST applied twice. The latter is currently the default for both the internal and the external vocabulary. All schemes have in common that a word can be decompressed individually and that there is one codebook per block of one million words (the block size can be configured).

This change requires a rebuild of the index.