Enable merging of `LocalVocab`s (#1310)
So far, operations with more than one operand (like joins) required that at most one operand had a non-empty local vocabulary. The reason was simply to avoid the need for merging two or more local vocabularies and the corresponding hassle of having to rewrite the local vocab indexes in one or more of the operand result tables.

Now local vocab IDs are actually pointers to 16-byte aligned strings (after removing the datatype bits, which cost 4 bits in precision, hence the 16-byte alignment). Each `LocalVocab` now maintains sets of such pointers and when two or more `LocalVocab`s are merged, we simply merge the respective sets. When comparing two local vocab IDs, the pointers are dereferenced.

This fixes #1242.