Use `LiteralOrIri` instead of `std::string` for expression results (#1318)
So far, intermediate results of expressions that were not `Id`s, were of type `std::string`. As a consequence, local vocab entries that were the results of expressions were missing the quotes in their internal representation. Now, all expression results have a proper type, namely `Id` or `LiteralOrIri`, and entries in the local vocab have a consistent representation.