Avoid re-sorting of index scan in transitive path operation (#1329)
A transitive path operation (like `?x wdt:P279* ?y`) first performs an index scan of the respective pedicate (`wdt:P279`) and then computes its transitive hull. So far, that index scan was alway ordered by subject. This led to a suboptimal query plan when the result of the transitive path operation is joined on the object (`?y`) with some other part of the query. Namely, the index scan was then re-ordered (using an expensive sort operation) before the transitive hull operation. This spurious sort operation is now avoided and the index scan with the right ordering is chosen in the first place.