Fix Docker badge in README.md (#1370)
The name of the Docker workflow has changed, thus the link to the corresponding badge in the `README.md` was broken. For some reason that badge was then always shown as "passed" although the workflow didn't exist anymore. This is now fixed and the badge should again become red on failure of the Docker build on the master branch.