SERVICE clauses now request JSON instead of TSV from the remote endpoint (#1361)
The RDF JSON format is a well-specified and lossless serialization of query results. This doesn't hold for th e TSV format, which
might lose information (e.g. about datatypes) and vary between different SPARQL engines.
With this change, more SERVICE queries work as expected, in particular when the result of the SERVICE contains literals that contain escape characters, non-ascii characters, or Literals with datatypes.

Current limitations:
* Blank nodes in the result of the SERVICE query are not yet supported, but as of this commit, this case will cause a human-readable error message that allows the user to work around this limitation.
* We currently have to store and parse the complete JSON response at once, which might take up quite some space. However, A lazy JSON parser as well as a lazy JSON exporter are currently being worked on, s.t. we can also consume the SERVICE response in chunks (as was the case for the old TSV implementation).