SPARQL Update Part 1: Located Triples (#1379)
This is the first PR in a series of PRs that will implement SPARQL UPDATE for QLever.

It implements `LocatedTriples`. These are triples that know to which block in a given permutation they belong to and whether they are to be inserted or deleted. They come with the functionality to find the corresponding blocks for a set of triples (to turn them into `LocatedTriples` as well as the functionality to merge a block from a permutation with all the `LocatedTriples` that belong to this block. The latter function is the core functionality to incorporate the results of an UPDATE into subsequent queries.