Fix a bug in the cache key of `DISTINCT` (#1403)
The cache key of a DISTINCT operation previously did not change when the subtree stayed the same, but the variables on which the DISTINCT was performed changed. For example, when first executing `SELECT DISTINCT ?x ?y WHERE { ?x wdt:P30 ?y }` and then `SELECT DISTINCT ?x WHERE { ?x wdt:P30 ? }` on Wikidata, the result for the second query would be mistakenly taken from the cache, although those queries have different results. This is now fixed by adding the column indices of the variables to the cache key.