implement functions `IRI()` / `URI()` (#1401)
This function takes a string and converts it to an IRI. If the input already is an IRI, then this function is the identity function.
Note: As QLever currently doesn't support the resolution of relative IRIs via BASE declarations, this limitation also holds for the IRI/URI function. so `IRI("hello")` will simply becom `<hello>`.