More relaxed parsing of IRIs
So far, our Turtle parse requires that IRIs are standard-conform. For
example, when an IRI contains a space, parsing stops. Even Apache Jena
(which implements the standard rather scrupulously) is more forgiving
there and just outputs a WARNing. This PR now does the same.

Specifically: first try strict (standard-conform) parsing. If that
fails, try more relaxed parsing (which currently also allows a space in
the IRI).