More relaxed parsing of IRIs (#1448)
So far, our Turtle parser (with `"ascii-prefixes-only": false` requires that IRIs are standard-conform, and halts with an error otherwise. But even Apache Jena (which implements the standard rather scrupulously) is more forgiving. For example, when the input format is Turtle and an IRI contains a space, Apache Jena just outputs a warning and not an error.

This change now does the same: It first tries strict parsing, and if that does not work, it outputs a warning and tries more relaxed parsing. The additionally allowed characters are space, backtick, or any one of `{}|^`. I checked that this does not slow down parsing.

NOTE: So far, spaces in IRIs didn't work with `"ascii-prefixes-only": true` either. That way, the PubChem data (which has a few IRIs with spaces), could not be parsed with either option. The reason that it didn't work with `"ascii-prefixes-only": true` either is that with this option, the parsing simply looks for the next `>`, however only as far as the next *space* or newline. This has now been changed to look as far as the next `<`, `>`, `"`, or newline. This excludes that any of these four characters occur inside of an IRI, but that is reasonable.