Add fourth column for graph name (#1337)
All six permutation pairs (SPO, SOP, OSP, OPS, PSO, POS) now have an additional column for the graph name. It is the fourth column (before the two columns for the patterns). For now, all internal triples get a special internal graph name, and all other triples belong to the default graph.

IMPORTANT: This changes the index format. It also slows down the indexing by about 15% (which is reasonable, given that the number of columns increased from 5 to 6).