Provide framework for generic lazily evaluated operation results (#1350)
QLever now supports lazy operations, that can produce their result in batches instead of fully materializing them. This PR implements the general infrastructure needed for lazy evaluation, and provides a lazy implementation of the `IndexScan` and `Filter` operation. This means that queries that consist of a single triple and one or more subsequent FILTER clauses will not materialize the full index scan.