Store internal IRIs like `ql:has-predicate` in the normal vocabulary (#1465)
So far, there was special code for internal IRIs like `ql:has-predicate` or `ql:default-graph`. Now there is special code only in the first stage of the indexing building, but once the partial vocabularies are merged, they have a name in the normal vocabulary just like all other entities. In particular, this fixes the problem that `ql:default-graph` was shown as `3` in the output.