Fix result for implicit `GROUP BY` without matches (#1389)
For an implicit `GROUP BY`, the standard dictates that there should be a result (a single line) even when there are no matches. The aggregate value then depends on the aggregate function: 0 for `COUNT`, `SUM`, `AVG`, undefined for `MIN` and `MAX`, empty string for `GROUP_CONCAT`. This is now properly implemented and tested. Here is an example query: https://qlever.cs.uni-freiburg.de/olympics/iq1oph .

Along the way, improve the documentation in `src/engine/sparqlExpressions/AggregateExpression.{h,cpp}` considerably, which makes the complex code there a bit easier to follow and extend.