Separate the effects of `send` and `LIMIT` again
https://github.com/ad-freiburg/qlever/pull/1355 changed the effect of
the QLever-specific `send` parameter by clamping the `LIMIT` of the
query accordingly, and adding one if none was there. This has the effect
that, in the QLever UI, the result size of all queries with a result
with more tha 100 rows was since shown as `100`.

This PR tries to revert this unwanted effect while keeping the good
stuff from #1355 (which unified the code for the various export formats
and fixed various inconsistencies). The behavior is now as follows:

1. The query is first processed according to the `LIMIT` specified in
   the query. In particular, if no `LIMIT` is specified, the full result
   is materialized internally.

2. At most `send` bindings are exported. In the QLever JSON, there are
   now two fields: `resultsize` (the size of the internally computed
   result) and `sent` (the number of bindings exported).

The idea is to add an option in the QLever UI for automatically adding a
`LIMIT' to each query. That option should be activated by default (in
order to save computation time), but a user should be able to deactivate
it in case they are interested in the full result size.

Future work: When one is only interested in the size of the full result,
that can als be computed without fully materializing it. That would save
RAM compared to the current approach.