Make `Filter` request its subresult lazily even if the filtered result is materialized (#1478)
Because `Filter` is an operation that can only reduce the result size (often drastically) there are plenty of occasions where the input of the filter is too large to be materialized, but the output of the filter can be materialized. As of this commit we thus always try to compute this input of a FILTER lazily.Because `Filter` is an operation that can only reduce the result size (often drastically) there are plenty of occasions where the input of the filter is too large to be materialized, but the output of the filter can be materialized. As of this commit we thus always try to compute this input of a FILTER lazily.