Improve `ScanSpecification` + fix empty index scan bugs (#1495)
The `ScanSpecification` class now has two additional members: a `LocalVocab` and a set of graphs. The `LocalVocab` enables that one or several of the `col0Id`, `col1Id`, `col2Id` can be new IRIs or literals (note that our comparators work fine even exactly the same IRI or literal is contained in different `LocalVocab`s). This is important for SPARQL Update. The set of graphs enables filtering of the scan result by one or severval graph IRIs.

Also convert `ScanSpecificationAsTripelComponent` to `ScanSpecification` already in the `IndexScan` class (where we already have an `Index` available to convert the `TripleComponent`s to `Id`s). In a future change, we might want to get rid of `ScanSpecificationAsTripelComponent` altogether and produce a `ScanSpecification` right away.

On the side, clean up the code and fix several bugs for empty results of index scan operations that only occur with lazy operations.