Add support for `FROM` and `GRAPH` with fixed IRI (#1445)
This continues work from #1337 (add a column with the graph name to the index), #1444 (add parser for N-Quads), and #1482 (add graph info to block metadata). Queries with `FROM` and/or `GRAPH` with a fixed IRI can now be processed. Processing queries with `FROM NAME` or `GRAPH` with a variable will be implemented in a future PR.