Make `RAND` function conform to standard
So far, we used our `FastRandomIntGenerator` to generate values for
`RAND`. However, the standard dictates that `RAND` returns an
`xsd:double` from the range [0, 1). We now do this using our
`RandomDoubleGenerator`. It is around six times slower than the previous
generator, but still fast enough (4.8 ns / number vs. 0.8 ns / number).

Also added a performance test to `RandomTest` to be able to check the
speed of a random number generator quickly.