Clamp `POINT` literals that are out of bound
So far, the index build exits when a `POINT` literal (of type
`geo:wktLiteral`) is encountered, where the latitude or longitude is out
of bounds. Unfortunately, Wikidata has several such points. This is a
simple way to address this: the latitude and longitude are clamped to
the permitted range and a warning is printed for each point for which
that happens.

Alternative solutions: drop the triple altogether or leave the literal
value as it is, but remove the datatype or change it to `xsd:string`.