Store the internal and external triples separately (#1532)
The internal triples (for the predicates `ql:has-pattern` and `ql:langtag` and all the `@lang@...` predicates) are now stored in a separate PSO&POS permutation pair and corresponding index files. This makes the handling of internal triples in the rest of the code much simpler. Also, internal IRIs now no longer leak for any query, unless it uses an internal IRI itself. In particular, this fixes #1513.

This is an index-breaking change. In particular, it adds new index files with names `<basename>.internal.<suffix>`.