Introduce nearest neighbor search for points (#1528)
QLever now supports a special predicate `?left <nearest-neighbors:k> ?right` where `k` is a positive integer that is transformed into a spatial join operation which for each entry in `?left` only returns the `k` spatially closest elements in `?right`. 
For the semantically correct usage, it is important that the `<nearest-neighbors>` join is performed as the last step, e.g. by putting the other constraints for `?left` and `right` into subqueries. In the future we probably want to replace this special predicate by a special construct that better reflects the non-commutativity of the nearest neighbor join.

Currently, the nearest neighbor join (same as the `<max-distance-in-meters>` join only work for point objects). The join is implemented using Google's `S2` library, delivers the functionality for nearest neighbor search out of the box.