Add function `geof:isWktPoint`
The function returns true if and only if the argument is a WKT point.
This can be checked efficiently by checking the datatype bits of the ID.

Note that this function is not part of the GeoSPARQL standard. We add it
because we need it for https://github.com/ad-freiburg/qlever-petrimaps
to fetch alll WKT literals that are not points efficiently.