Store the index of each block in the block's metadata (#1571)
This is a preparation for SPARQL UPDATE: When working with a subset of all blocks (e.g. because some prefiltering was applied), then we still have to know the absolute index of that block to find the corresponding `LocatedTriples`. The easiest way to find that index is to explicitly store it in the `CompressedBlockMetadata`.