Rewrite local vocab entries and blank nodes for `DeltaTriples` (#1580)
Each `DeltaTriples` object has exactly one local vocab and one blank node manager. When inserted triples contain local vocab IDs or blank node IDs, these are rewritten for that local vocab and blank node manager. Different local vocab entries referring to the same string are rewritten to the same new local vocab ID. Similarly, different occurrences of the same blank node are rewritten to the same new blank node ID. The rewriting also ensures that local vocab entries or blank nodes that were created merely during the evaluation of the WHERE clause of an insert operation, but do not end up in one of the resulting triples, do not become part of the local vocab or blank nodes of the `DeltaTriples` object.