Extend all index lookup methods to consider delta triples (#1582)
Extend the interfaces of all index lookup methods by either a `DeltaTriples` object (on levels, where the specific permutation is not yet known) or a `LocatedTriplesPerBlock` object (when the specific permutation is known). So far, they all receive an empty such object, which is then ignored. In a next step, the code will be added to pass an actual object and use it.