Prevent memory allocation errors because of caching for lazy results (#1578)
Lazy operations materialize their result up to a certain threshold to store it in the cache if it is sufficiently small. If these thresholds were configured strangely (in particular close to the total memory limit of the engine) it could happen, that a lazy operation failed with an out-of-memory error because of that caching mechanism. This commit fixes this unwanted behavior: As soon as the caching encounters an out-of-memorry error, it simply discards the materialized partial result for the chache, but the lazy execution continues unharmed.