Add a numPathsPerTarget parameter to PathSearch
Prior to this change, the PathSearch query used a lot of memory for
certain datasets and queries. In that case, the result can not be build
and it is not possible to determine where the problem lies.

This change adds the parameter numPathsPerTarget. The PathSearch will
only search and store paths if the number of paths is lower/equal to the
parameter value.