Allow unbound variables in compliance with the SPARQL standard (#1586)
So far, QLever reported an error when an unbound variable was used in a `GROUP BY`, `ORDER BY`, or at other places, where an unbound variable does not really make sense. However, the SPARQL standard allows this. For example, `SELECT * WHERE { ?s ?p ?o } GROUP BY ?doof` is allowed by the standard and the correct result is a table with a single column `?doof` but no results.

With this change, QLever complies to the standard regarding unbound variables at all kinds of places. Since queries like the above are often due to a typo in the query, we add a warning to the `application/sparql-results+qlever` results, which, in particular, is displayed in the QLever UI.

Additionally, there is a new runtime-parameter and associated command-line argument `--throw-on-unbound-variables` which reinstates the old behavior and reports an error instead of just a warning.

Fixes #1575