Allow block-based join algorithm to work with undef values (#1616)
Extend the block-based merge-join algorithm, s.t. it can support undefined inputs. This currently only works for single columns joins where the undefined values are at the beginning of each input and (as there is only one join column) compatible with every other input element. 
This is an important precondition for a lazy single-column join for sorted inputs.