Rename `LOG(...)` macro to `AD_LOG_...` to avoid conflicts with Abseil (#1629)
So far, QLever uses the macro `LOG(...)` for logging, where the argument is the log level. This is used by other libraries, too, in particular by Abseil. With an unfortunate include order, this leads to compilation errors that are hard to understand for new developers. We now use our usual prefix also for logging and call the macros `AD_LOG_INFO`, `AD_LOG_DEBUG`, etc.

---------

Co-authored-by: Hannah Bast <bast@cs.uni-freiburg.de>