Support joins between full index scans. (#1645)
So far, joins between full index scans were explicitly disallowed because their results are bound to be huge. Now that most operations are lazy, they make more send, for example, when counting or grouping the result. They are hence allowed now. In that context, allow `FILTER` to be applied as soon as possible also when a full index scan is involved.