Deduplicate the "other sets" in a `LocalVocab`  (#1632)
Since #1524, lazy joins can produce `LocalVocab`s with many duplicate "other sets". These duplicates cannot be easily detected because they are stored in a `std::vector`. They are now stored as a `absl:flat_hash_set` instead. This may come with a small performance penalty when the size of this set becomes large. To be able to detect this, add the size of the set to the details of the runtime information.

Based on the suggestion from https://github.com/ad-freiburg/qlever/pull/1524#discussion_r1851855712