First step towards a correct implementation of BASE
When a BASE IRI is defined, relative IRIs should be resolved against it.
For example, with `BASE <http://example.org/>`, the relative IRI `<a>`
should be resolved to `<http://example.org/a>`.

This is a fist step in this direction. So far, most of the test fail
because they use relative IRIs extensively, assuming that they are
left untouched.

PS: UniProt uses BASE extensively, so this is a blocker for parsing
UniProt via Turtle.