`BASE` declarations are now handled correctly (#1660)
When a `BASE` IRI is defined, IRIs without a scheme (like `<http://...>`) should be resolved. For example, with `BASE <http://purl.uniprot.org/uniprot/>`, the relative IRI `<UPI001AF4585D>` should be resolved to `<http://purl.uniprot.org/uniprot/UPI001AF4585D>`, and the absolute IRI  `</prosite/PS51927>` should be resolved to `<http://purl.uniprot.org/prosite/PS51927>`.

NOTE: Without `BASE` declaration, relative IRIs like `<a>` are accepted and left unchanged. This is used extensively in our tests. This is not 100% compliant with the SPARQL standard, but we consider being slightly more accepting than the standard harmless in this case.