`ql:contains-word` now can show the score of the word match in the respective text (#1397)
The fulltext index of QLever has forever been able to associate the occurence of a word in a text with a score. 
This PR adds the functionality to actually retrieve this score and to use it in the remainder of the query.
Currently the score is bound to a variable the name of which is automatically determined from the involved literals and variables. The easiest way to get the names of these variables is to use `SELECT *` or to look at the runtime information tree.