Add macros `QL_CONCEPT_OR_NOTHING` and `QL_CONCEPT_OR_TEMPLATE` (#1686)
The two macros can be used for concepts that improve the semantics and safety of the code, but are not necessary for compilation. They can be dropped or replaced by something simpler when compiling with C++17.

The macro `QL_CONCEPT_OR_NOTHING` can be used for concepts that can be omitted when compiling with C++17. For example, `QL_CONCEPT_OR_NOTHING(std::view) auto x = someFunction()`.

The macro `QL_CONCEPT_OR_TEMPLATE` can be used for concepts that can be replaced by `typename` when compiling with C++17. For example, `template <QL_CONCEPT_OR_TYPENAME(ql::same_as<int>) T> void f(){...}`.