Add option `--parser-buffer-size` for `IndexBuilderMain`
So far, the buffer size of the parser was hard-coded in
`src/index/ConstantsIndexBuilding.h` to `10 MiB`. This was too little
for datasets like OSM, which have some very large triples (due to WKT
literals of complex geometries). There is now an option
`--parser-buffer-size` with which the default buffer size can be
overridden.