Fix spatial join cost estimate
The cost estimate didn't include the cost of the children previously. This bug led to non-sense output in the runtime information JSON output in some cases, because the RuntimeInformation::getOperationCostEstimate subtracted child cost estimates that were larger than the spatial join cost. The resulting negative number interpreted as size_t (unsigned) caused an output near 2**64. This is now fixed by adding the child cost to the spatial join cost estimate.