Fix the Docker build on ARM (#1733)
The cross-compilation of the Docker image for ARM64  has been failing for a few days now, because a buggy QEMU version was downloaded by the used GitHub actions. We have now fixed QEMU to a working version. In the future we should build the ARM container natively and avoid the cross-compilation completely.