Also support `Authorization:Bearer` header to specify access token (#1720)
The access token (for privileged functionality) could so far only be passed as a parameter (that is, either as a field in the data for `application/x-www-form-urlencoded` or as a query parameter). With this change, the access token is also read from the `Authorization` header, provided it exists and the value has the form `Bearer <access token>`. If the access token is specified both via the `Authorization` header and via the `access-token` parameter, they must be the same. Resolves #1691