Reduce the memory overhead of the `augmentedMetadata` (#1734)
The current version of the code makes two copies of the complete block metadata, even when no updates are ever performed. For datasets like OpenStreetMap or UniProt this caused a base RAM consumption of 40 GB+ versus the usual 13 GB. The copies also noticeably slowed down the time needed for starting the server.

This is now fixed by storing the block metadata (for each permutation) as a shared pointer and adapting all the functions and interfaces depending on this. As long as there are no updates, the base RAM consumption of the server is now as it used to be (before update was implement). Once there is an update, the block meta data is copied and stored twice (original and augmented) from then on.