Don't use zero cost for cached subtrees
This continues #1736, which changed the size estimate for cached subtrees
from the exact size to the estimate size. Analogous to that, now also
change the cost estimate for cached subtrees from zero to the normal
cost estimate.

Introduce a runtime parameter `zero-cost-for-cached-subtree` to revert
to the old behavior if needed (the default is `false`).