Don't use zero cost estimate for cached subtrees (#1783)
This continues #1736, which changed the size estimate for cached subtrees from the exact size to the estimate size. Analogous to that, now also change the cost estimate for cached subtrees from zero to the normal cost estimate. That way, the cache no longer influences query planning (but can still improve query processing times because cached subtrees do not have to be computed again).

Introduce a runtime parameter `zero-cost-estimate-for-cached-subtree` to revert to the old behavior if desired (the default is `false`).