Implementation of `EXISTS` and `NOT EXISTS` (#1703)
This implements `EXISTS` via a new `ExistsJoin` operation. Namely, for each operation involving an expression (for example, a `BIND` or a `FILTER`), an `ExistJoin` operation is added for each occurence of `EXISTS` in that expression. The `ExistsJoin` adds an additional column (with a unique variable name) that contains the result of the `EXISTS` and is used to evaluate the expression. This can can be seen in the "Analysis" tree when executing a query. The current implementation has the following limitations, which will be addressed in future PRs:

1. The `ExistsJoin` operation is not yet lazy
2. The argument of an `EXISTS` is currently handled completely independent from the rest of the query (except for `FROM` AND `FROM NAMED` clauses, which are inherited from the outer query); it is an ongoing debate whether that is correct when the `EXISTS` contains `FILTER`s that refer to variables from outside the `EXISTS`.
3. When there are `UNDEF` values in the join columns, the `ExistsJoin` uses our generic zipper join, which is not particularly efficient