Change `tripleTemplate` rule to an equivalent non-recursive definition (#1816)
In the original `SPARQL` grammar, the `tripleTemplate` rule is defined recursively. This leads to stack overflows in the ANTLR-v4 based parser for large inputs, in particular large `INSERT DATA` requests. This PR changes the definition of the `triplesTemplate` to an equivalent formulation that is not recursive. This massively improves the performance of parsing large triple payloads. With this change we can process the whole olympics dataset as a single `UPDATE` request (1.78M triples, 323M).

To test this, `curl` with `--data-binary` can be used. Prepare the Update to execute into a file `foo`. Execute the query with `curl --data-binary @foo ...`. [Some more info on the different curl options](https://github.com/ad-freiburg/qlever/issues/1756#issuecomment-2665063376).

| Size    | Time before (s) | Time after (s) |
| ------- | --------------- | -------------- |
| 1781625 | N/A             | 65            |
| 800000  | N/A             | 27             |
| 200000  | N/A             | 7.5            |
| 100000  | N/A             | 4.2            |
| 92000   | N/A             | 4.0            |
| 88000   | N/A             | 3.8            |
| 84000   | 69              | 3.7            |
| 75000   | 59              | 3.4            |
| 50000   | 13              | 2.6            |
| 10000   | 1.7             | 1.3            |