Support post query values in top-level query (#1856)
Implement a `VALUES` clause that appears after the query body.
The `VALUES` clause is currently joined at the very end. In the future, we can create more optimized query plans that incorporate the clause into the query plan earlier, but there we have to be careful with the semantics of `FILTER`s with columns that contain unbound values.