Fix handling of unconnected `OPTIONAL`s (#1847)
`OPTIONAL` clauses that either appear at the beginning of the query or that are unconnected to the part of the query that stands before them behave either like an "ordinary" group graph pattern (if their result is non-empty) or like the neutral element of join (a single row with no bound variables) otherwise. So far, QLever only implemented the first case, leading to wrong results for unconnected `OPTIONAL`s that have 0 results. This is now fixed.
Fixes #1761