Don't apply language filter if no matching triple pattern exists (#1868)
QLever has two efficient ways of applying language filter: One using the tagged predicates like `@en@rdfs:label` and one that joins with the `ql:langtag` predicate. The latter one can only be safely applied, if the filtered variable appears somewhere in a triple, otherwise entities with a language tag that are created by the query (e.g. in a VALUES clause) are filtered out even if they have the correct language tag. This last restriction was missing so far and is introduced via this PR.

Fixes #1312