Fix inconsistency in hash and `operator==` for IDs with LocalVocabIndex (#1880)
For two Ids where one is a LocalVocabIndex and the other one is either a LocalVocabIndex or a VocabIndex,  `operator==` (and `operator<=>`) have a complex comparison logic that doesn't only look at the bits.
Before this PR, the hashing function only considered the bits, which lead to the case that there were IDs that had different hash values, but compared equal, which is a serious bug that breaks hash sets and hash maps for these IDs.
This PR fixes that bug by incorporating all the data of an ID into the hash that is also relevant for the equality comparison.
Fixes #1863