Support BM25 and tf-idf scores in the full-text index (#1688)
With this PR, there are now three different types of scores for a word occurence in a document in the full-text index. The score has to be chosen when the text index is built. The scores are as follows:
* Explicitly defined scores in the wordsfile (the only option so far)
* tf-idf scores
* BM25 scores, with configurable parameters `b` and `k`.
Note that the tf-idf and BM25 scores take into consideration both the words- and the docsfile, so they currently have the following behavior:

Only words from the wordsfile become part of the text index, but the occurences of these words  in the docsfile have to be counted to compute the scores. This means that the score values will be wrong (too low) if the words as specified in the wordsfile are not found in the documents because the default tokenizer in QLever is inconsistent with the one used to declare the wordsfile.

In the future we will implement a configurable tokenizer for the docsfile, to make it thus easier to mitigate the above caveat or to get rid of the explicit wordsfile at all.