Correct datatype and language tag for the result of `SUBSTR` (#1636)
If the input string of the `SUBSTR` function has a language tag, then according to the SPARQL standard the result has to preserve this language tag. Same goes if the datatype of the string is `xsd:string`. This PR implements this behavior (previously the language tag and datatype were always dropped) and implements all the infrastructure to implement this behavior also for other expressions that take and return strings.om>