Correctly handle dataset clauses parsed as HTTP parameters (#1845)
According to the SPARQL Protocol standard, dataset clauses that are specified as HTTP query parameters override all `FROM [NAMED]` in the given query or update request. This PR implements this behavior by extending the SPARQL parser with the option of setting fixed datasets that cannot be overwritten during the parsing, even when a FROM clause is encountered.

Also refactor several tests of the SPARQL parser, in particular by consistently separating the actual test code from generic helpers like gmock matchers.