Correctly handle SPARQL Update requests with graph variables (#1860)
Fix the handling of Graph variables inside Updates. An example that now works is the following query that deletes the matching triples from all graphs (the unnamed default graph and all named graphs):
```sparql
DELETE WHERE {
  GRAPH ?g {
    ?s rdfs:label ?o
  }
}
```

As a part of this fix, the internal representation of Update clauses was refactored to properly handle quads of all kinds (quads with a fixed graph, a variable graph, and the implicit default graph).