Two helper ranges to simplify the rewriting of co-routines to C++17 (#1892)
This introduces `CachingTransformInputRange` and `CachingContinuableTransformInputRange`, two abstractions that will make the porting from C++20 generators to C++17 easier.

`CachingTransformInputRange` is similar to `ranges::transform`, but caches the intermediate results by value and always is an `InputRange`. This is closer to the behavior of coroutines than `ranges::transform` which in particular recomputes the transformation if the same iterator is dereferenced multiple times.

`CachingContinuableTransformInputRange` in addition has special return values for the transformation which can control the iteration, e.g. `Continue` (skip the current element).