Persist updates between server restarts (#1903)
So far, delta triples due to updates were lost when the server was restarted. This change introduces an option `--persist-updates`, which has the following effect:

1. After each update, the complete set of delta triples (inserted and deleted) as well as the associated local vocabulary is written to disk, to a file `<name>.update-triples`, where `<name>` is the basename of the index files.

2. When restarting the server, that file is read and the update triples are processed in such a way that the server ends up in the same state it was in when that file was written.

3. When restarting the server without the option `--persist-updates` or if there is no file `<name>.update-triples`, the server is in the same state it would be in if there are no update triples (or after the command `clear-delta-triples`).

NOTE: This is a first implementation, which provides the basic functionality. After each update, the complete set of delta triples is written, even if that update inserted or removed only few triples relative to the previous updates. If an error occurs during an update operation, the server state is still undefined and one effectively loses all delta triples. This is work for future PRs.

Co-authored-by: Johannes Kalmbach <kalmbach@cs.uni-freiburg.de>
Co-authored-by: Hannah Bast <bast@cs.uni-freiburg.de>