Implement the built-in function `BNODE` (#1911)
The function `BNODE` can be called without argument or with one argument. Without argument, `BNODE()` returns a unique blank node. In particular, different function calls return different blank nodes. With an argument, `BNODE(...)` returns a blank node that is different from all other blank nodes but the the same for different function calls of `BNODE(...)` with the same argument. Internally, the blank nodes of the latter type are represented as special IRIs. This works but is not memory-efficient when there are many such blank nodes (this is work for a future PR). Here are two example queries for testing that the function does what it should do:

```sparql
SELECT (BNODE("1") AS ?a) (BNODE(?x) AS ?b) (BNODE() AS ?c) (BNODE() AS ?d) WHERE {
  VALUES ?x { 1 2 3 }
}
```

```sparql
CONSTRUCT {
  ?x ?y []
} WHERE {
  VALUES ?x { 1 2 3 }
  BIND (BNODE() AS ?y)
}
```

Co-authored-by: Johannes Kalmbach <kalmbach@cs.uni-freiburg.de>