Add pre-filter for `isBlank`, `isNumeric`, `isLiteral` and `isIri` + Optimize `PrefilterExpression::evaluate` (#1876)
With this PR, queries like
```
SELECT ?o WHERE {
  ?s ?p ?o
  FILTER ISBLANK(?o)
 }
```
Become more efficient, because only the relevant blocks are read in the index scan. 
Also the whole evaluation if these prefilters is optimized by not copying the (rather large) `CompressedBlockMetadata` in every intermediate step, but only at the end.
Note that this optimization can be propagated further in the future, because the result of the prefiltering always is a subset of all blocks that is cheap to encode.