Make `<libspatialjoin>` aware of WKT geometries from `LocalVocab`  (#1941)
When doing spatial joins with `<libspatialjoin>`, the `LocalVocab`s of the two input tables were ignored so far. In particular, geometries that were not part of the original index were therefore ignored. This is now fixed. In particular, fixes #1939

On the side, drop `--ffast-math` from the compilation options of `libspatialjoin` to fix some warnings in the `fmt` library and `Geo.h`, in which `infinity` was still used