Phase out `STXXL`, part 2 = now completed (#1940)
This completes the phase-out of `STXXL`, started in #1926, by replacing the few remaining uses with already existing functionality like `ad_utility::CompressedExternalIdTableSorter`. Background: https://github.com/stxxl/stxxl is no longer maintained. As of this writing, the last commit is from 15.11.2018. The text index build is slightly faster now (about 10% for Wikidata with the English Wikipedia as text).