Parallel Turtle parsing, better error messages for scattered base and prefix declarations (#1871)
The parallel Turtle parser requires that all `PREFIX` and `BASE` declarations come in a single block at the beginning of the file. For declarations after this initial block, confusing things could and often did happen. It could happen that the later declaration was ignored for some triples (which is wrong if the declaration changed a previously defined IRI). Or a confusing error message was given (that the prefix was undeclared).

This is now fixed. If a declaration is repeated but does not change the definition from the initial block, it is silently ignored (which is correct). If a declaration is repeated but does change the definition, a proper and detailed error message is provided. Fixes #1794.

NOTE: This is an improved version of #1807, which was merged on 18.02.2025 but then reverted again because it broke the Wikidata build (where declarations are repeated again and again but without changing the IRI).