Fix `SERVICE` request with no variables (#1975)
For a `SERVICE` request without visible variables (in particular, for an "empty" request like `SERVICE <iri> { }`), so far QLever sent an invalid query of the form `SELECT WHERE { ... }` to the respective endpoint. This special case is now fixed by sending `SELECT * { ... }` instead. The `WHERE` is now omitted, also for the general case, which saves six bytes. The unit tests were improved to provide better line numbers if they fail. Fixes #1765