Add `CONTRIBUTING` guidelines + draft for `NOTICE` file
Now that more people from outside of the Chair of Algorithms and Data
Structures are contributing, we need guidelines on how to acknowledge
those contributions in the individual files. This is now explained in
a file `CONTRIBUTING` in the root directory.

Also, the `NOTICE` file hasn't been updated for a very long time.
According to the conditions of the Apache License, it needs to acknowledge
the third-party software used in the project, according to the conditions
of their respective licenses.