Give `ProgressBar` users control over timer if needed
So far, the timer for a `ProgressBar` was always started as soon as the
`ProgressBar` was created. This distorts statistics for applications
with a large initial wait time. For example, this happens during index
building when one of the input streams is not just produced from a file
but computed (e.g., using `osm2rdf` for the OSM data).

The `ProgressBar` class now has a `getTimer()` method that gives the
user more control. This is used in `IndexImpl::passFileForVocabulary()`
to start timing only once the first triple has been parsed.