Add git hash to result for `?cmd=stats` request (#2004)
Add two fields `git-hash-index` and `git-hash-server` to the result for a `?cmd=stats` request. This makes it easy to find out the version of the server and of the index builder (at the time of the index build) for a particular QLever endpoint. In particular, this information can then be displayed in the QLever UI. For now, we make this information available without access token, but this could easily be changed in the future if desired or needed.