Add GCC-8 C++17 build with error tolerance (#1764)
This PR adds a CI check for GCC-8 in C++17 mode. As this check currently still fails consistently, this build is error-tolerant but allows us to track the progress of the migration to C++17. The build supports specifying some targets (e.g. unit tests) that are expected to compile in C++17 mode and will make the build fail on error.