Lazy `OPTIONAL` join for single join column (#1996)
So far, the processing of `OPTIONAL` always materialized the result of both of its input group graph patterns, as well as the joined result. In particular, queries with the rather typical ` ... OPTIONAL { ?subject <some_predicate> ?label }` failed when `<some_predicate>` or the result of `...` were large.

This is now fixed for the most common use case when there is a single join column (that is, when there is exactly one common variable between the left-hand side and the right-hand side of the `OPTIONAL`), and at least one of the two sides is provided lazily. Then the output is provided lazily, too. The join column may contain `UNDEF` values.

NOTE: When a now lazy `OPTIONAL` join is followed by a `GROUP BY` with non-trivial aggregates (like `SAMPLE`), queries are now significantly slower. This is not the fault of the new lazy implementation of `OPTIONAL`, but due to the fact that `GROUP BY` with non-trivial aggregates is much slower for lazy inputs than for materialized inputs. This will be addressed in a separate PR.