Make `<=>` for `LiteralOrIri` respect language tag (#2035)
So far, there was a subtle inconsistency between the `==` and `<=>` operators for `LiteralOrIri`. For example, `"foo"@en` and `"foo"@de` were not equal according to `==` (which has the default implementation), but equal according to `<=>` (which has a custom implementation using `staticSingletonComparator`). The reason is that the latter ignored the language tag, because the level defaults to `QUARTERNARY`. The level is now explicitly set to `TOTAL` for the comparison, which fixes the inconsistency. Fixes #1990