Support chained updates (#1883)
SPARQL 1.1 Update allows multiple operations in a single request, separated by a semicolon. These operations then have to be applied atomically. This is now implemented.

NOTE: The implementation is currently not as efficient as it could be. In particular, a new snapshot of the located triples is made after each operation instead of once for the whole request. This should be addressed in a later PR: