Add pre-filter for `REGEX` and `STRSTARTS` (#1985)
This makes combinations of an index scan and a filter as in `?person rdfs:label ?label FILTER STRSTARTS(?label, "Barack")` or `?person rdfs:label ?label FILTER REGEX(?label, "^Barack")` very efficient. Namely, the index scan reads and decompresses only those blocks from the respective index list, which according to their metadata (in particular, the first and last triple of the block) potentially match the filter condition. This continues work from #1876 and previous PRs of this kind.