Fix `LIMIT` and `OFFSET` for subqueries (#2060)
So far, QLever did, in certain cases, not correctly apply the `LIMIT` to subqueries. In particular, this happened for `SELECT * { SELECT * { ?s ?p ?o } LIMIT 1 }`. This is now fixed. In particular, fixes #2013.

The previous `Operation::setLimit` method is now called `Operation::applyLimitOffset` and considers a previously set limit and offset instead of just replacing it. Also, there is now a new method `Operation::onLimitOffsetChanged`, which should be a no-op for operations for which `Operation::supportsLimitOffset` returns `false`, but can otherwise be overridden to implement limit and offset more efficiently.