Fix Docker build on macOS (`signed char` vs. `unsigned char` issue) (#2085)
In `TextIndexBuilder::fourLetterPrefixes()`, there was a call to `views::iota('a', 'z' + 1)`, which works fine when `char` is `signed char`  (which is true for most systems), but produces a warning when `char` is `unsigned char` (as is the case for our macOS build). This is now fixed by explicitly writing `views::iota(char{'a'}, char{'z' + 1})`. On the side, simplify the code by using `views::cartesian_product`.