Add `PolymorphicVocabulary` class with unit tests (#2070)
Enable a choice (at runtime) between four kinds of vocabularies: `InMemoryUncompressed`, `OnDiskUncompressed`, `InMemoryCompressed`, and `OnDiskCompressed`. The only restriction is that the vocabulary must be read from disk with the same choice that was used for writing it.

This change adds the new class `PolymorphicVocabulary` together with some unit tests. This prepares #1740, which will integrate this class into QLever.