Return `Result::idTables()` by value (#2110)
Previously, `Result::idTables()` returned a `LazyResult` by reference, which could be used only once. The function was often used as a temporary, as in `other_function(std::move(res->idTables(), ...)`. This did not make sense and was misleading. Now the `LazyResult` is returned by value and the function is used without `std::move` when used as a temporary.