Support uncompressed or fully in-memory vocabularies (#1740)
So far, QLever's vocabulary (the set of distinct IRIs and literals) is compressed (on the level of individual IRIs and literals) and resides on disk by default. It can be configured that certain IRIs or literals reside in RAM. However, there was some overhead for determining whether an IRI or literal resides in memory or on disk. With this change, `IndexBuilderMain` gets a new option `--vocabulary-type` with the following five choices:

`on-disk-compressed` (the default so far, and still the default)
`on-disk-uncompressed` (costs more space but saves time for decompression)
`in-memory-compressed` (faster than disk but memory is expensive)
`in-memory-uncompressed` (maximally fast but most memory hungry)
`on-disk-compressed-geo-split` (like the default, but with WKT literals in an own geo vocabulary)

Here are performance results for building an index for https://dblp.org/rdf/dblp.ttl.gz (514 M triples) and then exporting a TSV with 100 M triples (total size 9.2 GB) on a Ryzen 9 9950X with 4 x 8 TB (SN850X NVMe, PCI-E 4, RAID 0) and on a Ryzen 9 5900X with 1 x 4 TB (SN850X, PCI-E 4). Each measurement ist taken after clearing the disk cache (with `sudo bash -c "sync; sleep 5; echo 3 > /proc/sys/vm/drop_caches"`) and restarting QLever (with `qlever start`). When repeating the export (without restarting QLever, but after `qlever clear-cache`), the numbers are not significantly different. The first column names the total size of the vocabulary files on disk (for the in-memory options, the vocabulary is read from disk when the server starts).

| on-disk-compressed | 2.0 GB | 72s | 179s |
| on-disk-uncompressed | 3.9 GB | 63s | 160s |
| in-memory-compressed | 2.0 GB | 32s | 50s |
| in-memory-uncompressed | 3.9 GB | 24s | 35s |

Co-authored-by: Hannah Bast <bast@cs.uni-freiburg.de>