Choose media type depending on the query type (#2133)
So far, the choice of media type did not respect the query type. In particular, the default media type (when no accept header was specified) was always `application/sparql-results+json`, which led to an error message for `CONSTRUCT` and `DESCRIBE` queries.

This is now fixed and made compliant with the standard. The media type is now the best fit for both the accept header (if any) and the query type. Note that an accept header can also specify multiple media types (separated by comma), wildcards (e.g., `text/*`), and priorities (e.g., `text/turtle;q=0.9, text/csv;q=0.1`).