Pad negative years in dates to 4 digits (#2160)
So far, negative years in dates where only padded to three digits (because the string formatter considered the leading `-` as a digit). This is now fixed: all dates with years in the range -9999..-1 will be exported as `-YYYY-...`, and all dates in the range 0..9999 will be exported as `YYYY-...`. In particular, fixes #2106