Enable substitution of `FILTER` clauses in query planner (#2119)
Enable the query planner to replace a `FILTER` clause (e.g., `FILTER(?a = ?b)`) with an equivalent subtree plan (e.g., `?a <equal-to> ?b`). This does not yet rewrite any queries, but prepares #1935, where this new mechanism is used to rewrite `FILTER` clauses that are equivalent to certain kinds of spatial joins.