In chained updates, make a snapshot before each update (#2118)
In a chained update, we have to make a snapshot of the located triples before each update. Otherwise, the query planner might conclude (before executing the query) that the result is empty when actually it is not due to a preceding update in the chain. Also, making a snapshot improves the size estimate and hence the query plan.

NOTE: Making snapshots is expensive, but we leave that aspect to a separate PR.