Add `GeoVocabulary` to store precomputed geometry data along with WKT literals (#1951)
When building an index with `--vocabulary-type on-disk-compressed-geo-split`, QLever now stores WKT literals using a new vocabulary class `GeoVocabulary`. Along with each WKT literal, a `GeometryInfo` object is computed and stored on disk. Each such object currently holds the geometry type (e.g., `"http://www.opengis.net/ont/sf#Polygon"^^<http://www.w3.org/2001/XMLSchema#anyURI>`), the centroid (a point), and the axis-parallel bounding box (a polygon with four points). Each object is represented efficiently in 24 bytes (8 bytes for the centroid and geometry bytes, and another 16 bytes for two points defining the bounding box).

With such an index, each of the three functions `geof:geometryType`, `geof:centroid`, and `geof:envelope` can now be computed much more efficiently, via a simple random lookup of the respective `GeometryInfo` object of disk for each WKT literal. Without such an index, the full WKT literal has to read from disk, which is much more expensive, especially for large WKT literals.

NOTE: This PR builds on #1740, #2111, #1983, #1961, and #2158. The `GeometryInfo` class is designed to be easily extensible. We might also want to add the ability to cache parts of it in RAM for faster access.