Allow prefix search in text for arbitrary prefix lengths (#2181)
So far, the minimal prefix length for prefix search using `ql:contains-word` was four. That is, `?text ql:contains-word "comp*"` was possible, but `ql:contains-word "com*"` was not. Now, an arbitrary prefix length is possible (even `"*"`, though not particularly useful), and if multiple blocks have to be read from the inverted index, they are simply concatenated and then sorted.

NOTE: This works, but is not yet as efficient as it could be when multiple blocks have to be read. Possible improvements are:
* Split the blocks up more efficiently (not by prefixes of a fixed length, but such that blocks have similar sizes)
* Implement a k-way merge of the presorted blocks from the half-inverted index.
* Make the merging lazy such that the memory consumption is bounded, even when a large amount of data has to be read

Co-authored-by: Johannes Kalmbach <johannes.kalmbach@gmail.com>