Fix Unicode support for text search with `ql:contains-word` (#2215)
Text search using QLever's special `ql:contains-word` predicate did not work correctly with non-ASCII characters because the text tokenization incorrectly split words containing non-ASCII characters and because the automatic variables for scores had mangled names when non-ASCII characters were involved. This is now fixed.

NOTE: This fix was started in #2176 by @wjbmattingly (thanks a lot); see that PR for a more detailed description of the problem (with examples) and the solution.