First steps to fix a bug in `CompressedExternalIdTable`
The bug occurs when building an index for a dataset, where all subjects
are blank nodes. Then the following assertion fails:

`Assertion `(numBlocksPushed_ == 0) != compressAndWriteFuture_.valid()`

The problem only occurs, when `STXXL_MEMORY` is not large enough.
However, when making it too small, the index build fails even earlier
with a different (already known) problem, namely that the setting is not
large enough for merging.