Improve the performance of `getBlocksForJoin` (#2252)
So far, `CompressedRelationReader::getBlocksForJoin` did a binary search on the join column for each block, that is, it had a complexity of `O(B x log J)`, where `B` is the number of blocks and `J` is the number of rows in the join column. This has now been improved to a zipper-like merge, which has a complexity of `O(B + J)`, at least when there is no significant Cartesian blow-up. This is significantly faster when both `B` and `J` are large. For example, it makes the last join in the following (artificial) query on Wikidata three times faster:
```sparql
SELECT * WHERE {
  ?s rdf:type ?o1 .
  ?s wdt:P31 ?o2 .
  ?s wdt:P31 ?o3 .
}
```

Co-authored-by: RobinTF <83676088+RobinTF@users.noreply.github.com>