Add new `StripColumns` operation (#2253)
Add a new `StripColumns` operation, which has a single child and keeps only the given subset of columns. This has a cost but saves computation further up in the query execution tree.

NOTE: For now, this is just implemented and tested but not employed by the query planner. It will be employed in subsequent PRs. Eventually, the various operations should be implemented such that they only produce those columns needed by subsequent operations.