Recursively strip columns in subqueries and for `GroupBy` (#2266)
When a subquery or `GroupBy` operation selects only a subset of the visible variables, this information is propagated recursively to the child operations. Each operation informed this way can then produce only the requested columns. For now, the following operations are able to consider and propagate this information: `Join`, `OptionalJoin`, `Minus`, `Sort`, `Union`. 

Introduce a new runtime parameter `strip-columns`. For now, it defaults to `false`, so that all the unit tests unrelated to this change (which assume unstripped columns) still work. Only in `ServerMain` set it to `true`, so that column-stripping is the default behavior when starting the server.

NOTE: When an operation is not in this list (for example, `Filter`), the propagation stops at the respective node in the query execution tree. For `IndexScan` operations, columns are stripped after the columns are read from the index. Information from the top-level select query is currently not propagated when there is no `GROUP BY`. These and other  issues will be addressed in future PRs.