Add runtime parameter `spatial-join-prefilter-max-size` (#2242)
With #2205, `SpatialJoin` can prefilter the geometries of the larger side of the join using the aggregated bounding box of the geometries on the smaller side. This potentially saves reading geometries from disk at the cost of reading all the bounding boxes first. If the aggregated bounding box is large, the cost can outweigh the savings. The current code uses the prefilter when the size aggregated bounding box is below or at a rather arbitrarily fixed threshold, namely `2,500` square coordinates.

There is now a runtime parameter `spatial-join-prefilter-max-size` to set this threshold value dynamically. The unit is still square coordinates and the default is still `2,500`. Setting the threshold to `0` effectively disables prefiltering.