Consider `GRAPH` variable in `TransitivePath` data structures (#2274)
So far, transitive paths (property + `?` or `+` or `*`) are not yet supported inside a `GRAPH ?g` clause. This is preparation for supporting this, by making the graph variable or its values available to the methods that need it. Add unit tests that test this functionality. Does not yet add the functionality to QLever, this will be done in #2128