Fix issue and add regression test (#2288)
The following query leads to an assertion error. This is a regression introduced by #2264. It is now fixed again. In particular, fixes #2284.
```sparql
SELECT (COUNT(DISTINCT ?g) AS ?count) WHERE {
  GRAPH ?g {
    ?s ?p ?o .
  }
}
```
NOTE: The query materializes all quads (triples + graph info) from the data, and thus only works when enough memory for that is available. Fixing this is a separate issue