fb w/o sample: Fix
I tried to export the creation of "ArrayOrVector<Id>" keys to a function but I failed. The function required a lambda, because the usecases are too different, it was ugly. Until I find a better solution, it may look like code duplication (though it may actually not be that way)