Fix prefilter for `isLiteral()` (#2300)
Since #1876, index scans in conjunction with `isLiteral` can be pre-filtered. However, this optimization falsely filtered out blocks with only bools, ints, and doubles, which are literals too. In particular, the result for `?count_yes` in the following query was wrong so far. This is now fixed, and the results for `?count_yes` and `?count_no` correctly sum to the result for `?cout_all` https://qlever.cs.uni-freiburg.de/wikidata/QQfbH7
```sparql
SELECT * WHERE {
  { SELECT (COUNT(*) AS ?count_yes) { ?s ?p ?o . FILTER isLiteral(?o) } }
  { SELECT (COUNT(*) AS ?count_no) { ?s ?p ?o . FILTER (!isLiteral(?o)) } }
  { SELECT (COUNT(*) AS ?tcount_all) { ?s ?p ?o . } }
}
```