Encode IRIs with a common prefix and integer suffix directly in the `Id` (#2299)
The prefixes can be specified with a new option to `IndexBuilderMain`, called `--encode-as-id`. For example, with `--encode-as-id http://example.org/`, all IRIs with that prefix and an integer suffix are encoded directly in the `Id`. For multiple prefixes (up to 256), use the option only once and separate the prefixes by a space; see below for an example of this.

The details of the encoding are described in a long comment in `src/index/EncodedIriManager`. Any IRI that does not match any of the specified prefixes (by default, the list of such prefixes is empty) or with an integer suffix with more than 13 digits, is stored like before, with a `VocabIndex` encoded in the `Id` that can be used to lookup the string representation in the vocabulary. In contrast, encoded IRIs do not require an entry in the vocabulary, which saves space and allows faster materialization of the string representation.

LIMITATIONS: The ordering of IRIs is correct among the encoded IRIs as well as among the non-encoded IRIs, however, not necessarily between them. For example, this may give unexpected results for `FILTER` or `ORDER BY` when IRIs of both kinds are involved. These limitations will be addressed in a future PR. If these limitations are critical, simply don't use the `--encode-as-id` feature.

PERFORMANCE: Here are the results of a small performance evaluation on the OpenStreetMap data of Switzerland (with a total number of 841,903,678 triples), using `--encode-as-id https://www.openstreetmap.org/node/ https://www.openstreetmap.org/way/ https://www.openstreetmap.org/relation/ https://osm2rdf.cs.uni-freiburg.de/rdf/geom#osm_node_ https://osm2rdf.cs.uni-freiburg.de/rdf/geom#osm_way_ https://osm2rdf.cs.uni-freiburg.de/rdf/geom#osm_relation_ https://www.openstreetmap.org/changeset/`.

1. The size of `.vocabulary.codebooks`, `vocabulary.words.external`, and `.vocabulary.words.internal` becomes 7% smaller
2. The size of each of the `.index.???` files becomes around 5% larger (because the `Id`s of the encoded IRIs are slightly less compressible since the gaps between IRIs with consecutive integer suffixes become larger in the encoding)
3. The time for exporting the full result of `SELECT * WHERE { ?s ogc:sfIntersects ?o }` (148,672,077 triples, each of which contains two encoded IRIs and always the same non-encoded IRI for the predicate) reduces from 76s to 19s, so very significantly
4. The time for exporting the full result of `SELECT * WHERE { ?s ?p ?o }` (841,903,678 triples, only a fraction of which contain encoded IRIs) reduces from 574s to 420s

Co-authored-by: Hannah Bast <bast@cs.uni-freiburg.de>