fb w/o sample: Add more tests, Fixes
Altered makeProcessGroupsVisitor (we have to skip nonmatching entries there when switching to the hybrid fallback).
We skip the current block that we have just processed.
For "inWidth" we use _subtree->getResultWidth() instead of "beginIdTable.numColumns()" because it caused "bad optional access" errors.
Removed argument annotations. The IDE can display them.