Fix bugs regarding identical triples contained in multiple graphs (#2090)
Fix several bugs for queries that didn't have a graph column, but identical triples were contained in multiple graphs. In particular:

1. Fix the graph duplicate filtering inside the `readPossiblyIncompleteBlock` function
2. The ordering of blocks was inconsistent in the presence of such duplicate triples, this is now fixed
3. The index builder now ensures that all triples that are duplicates regarding the SPO part (that is, without the graph info) are stored in the same block; that way, duplicate filtering never has to be performed across blocks (which would have been necessary so far, but was not done)
4.  In particular, when located triples are equal regarding the SPO part (that is, without the graph info) to the first or last triple of that block, then they always belong to this block (so far, they sometimes belonged to the next block)

On the side, small refactorings in the `CompressedRelation.cpp` class, in particular, the removal of unnecessary `shared_ptr`s. Some test cases in `LocatedTriplesTest` had to be rewritten because they worked with located triples that violated some of the preconditions that are now checked via `AD_CONTRACT_CHECK`s

Fixes #2134
Fixes #2002
Fixes #2339