Implement Graph Store HTTP Protocol operations `PUT`, `DELETE`, and `TSOP` (#2117)
This continues work from #1727, which implements the Graph Store HTTP Protocol operations `GET` (get all triples from the specified graph) and `POST` (insert triples from payload in the specified graph). This change also implements `PUT` (replace the triples in the specified graph with the triples from the payload) and `DELETE` (delete all triples in the specified graph).

Also implement the non-standard `TSOP`, which *deletes* the triples from the payload in the specified graph (TSOP is POST backwards). In particular, this is useful for processing the Wikidata update stream, where the to-be-added and the to-be-deleted triples are both provided as RDF Turtle and not as SPARQL updates operations.

Improve the behavior of `POST` by returning HTTP status code 204 when the request body is empty.

NOTE: `PUT` and `DELETE` are not particularly efficient yet, when the graph contains many triples previously to the call. And `TSOP` does not delete triples involving blank nodes, for the same reason that blank nodes are not allowed in `DELETE DATA`. This is fine for the Wikidata update stream, which does not contain blank nodes (for exactly that reason).