Fix expression evaluation for grouped variables in lazy `GROUP BY` (#2373)
When evaluating expressions on grouped variables, QLever substitutes, for each group, the `VariableExpression` with the grouped variable by a `VectorIdExpression` with the variable value (repeated x times if the group has x elements) . This breaks when the `GROUP BY` is lazy and more than one block is produced. This results in an "Assertion `numItems == vector.size()` failed" error. This change fixes the bug for most, but still not all expression. In particular, fixes #2349

NOTE: See #2374 for more details, including queries, where it still does not work as it should.