Vocabulary merging without `mmap`ed files (#2356)
In the first stage of index building, QLever parses the input in batches and creates a so-called partial vocabulary for each batch. Each partial vocabulary contains a list of distinct words and the preliminary ("local") `Id` for each word. After parsing, the partial vocabularies are merged. In the process of this merge, a map is created for each partial vocabulary that maps each "local" `Id` to the final "global" `Id` of each word. These maps are then used to produce a stream of `Id` quads for the various permutations. So far, these maps were written and read to disk using memory-mapped files. This turns out to be problematic when there are very many of them (e.g., more than 10'000 for UniProt), leading to confusing and annoying error messages.

These maps are now realized using ordinary files. While refactoring the code, give better names to various of the associated variables and the two kinds of files for the partial vocabularies (one kind for the word-local-id pairs, the other kind for the local-id-global-id map). Fixes #1966, fixes #2309, and fixes #1858.

NOTE on performance: It works fine on the latest version of UniProt from 2025-06-17 (294 B triples parsed in 11'765 batches, 64 B distinct words). There was no notable difference in the performance of the vocabulary merging phase compared to the previous implementation.

Co-authored-by: Johannes Kalmbach <johannes.kalmbach@gmail.com>