Fine-grained timings for each update operation (#2029)
Since #1675 each update operation returns a JSON with fine-grained statistics on the number of inserted and deleted triples (before, after, and by the operation), the number of index blocks affected, etc. This JSON also contains timing information, but not very fine-grained and the times did not quite add up to the total time for the operation.

This is now fixed. In particular, the following are now measured separately: time for query planning, time for preparation, time for handling insertions and deletions, time for creating the snapshot, time for writing the snapshot. Several of these have detailed sub-timings. This is realized via a new class `TimeTracer` that can also be used for the time-grained timing of other operations in the future.

This new timing information is already used by the `qlever update-wikidata` command; see https://github.com/ad-freiburg/qlever-control/pull/158.

NOTE: The time for parsing is not yet accounted for. Some of the names of the timings could still be improved. We leave that to a separate PR.