Fix export and `STR` result for doubles and decimals with integer values (#2358)
So far, a double or decimal with an integer value of, say, 12345, was exported as, `12345`. This is not standard-compliant and for a good reason: When reading such a result back in (e.g., from `text/turtle` generated by a `CONSTRUCT` query), the value would become an integer.

The value is now exported as `12345.0`. However, when converted to a string, either implicitly or using `STR`, the standard demands that the string is `12345` without the decimal point. This is now ensured as well.

NOTE: The standard has further requirements on the conversion of a double to a string, see https://www.w3.org/TR/xpath-functions/#casting-to-string . In particular, the standard demands that values between `0.000001` (one millionth) and `1000000` (one million) are converted to a string like a decimal, while for other values (except for `0`, `-0`, `INF`, and `-INF`), the `E` notation must be used. This is not yet implemented and work for a separate PR.