Make configurable whether `GRAPH ?g` matches the default graph (#2352)
According to the SPARQL standard, `GRAPH ?g` must not match the default graph. This is now fixed. In particular, the following query now returns the size for each named graph, but no longer for the default graph. This behavior can be controlled by the new runtime parameter `treat-default-graph-as-named-graph`. When `false`, which is the default, the default graph is not considered a named graph (which is what the standard demands). When `true`, it is (which is the previous and seemingly more natural behavior).
```sparql
SELECT ?g (COUNT(?g) AS ?count) {
  GRAPH ?g {
    ?s ?p ?o
  }
}
GROUP BY ?graph
ORDER BY DESC(?count)
```