Rebuild the index permutations PSO and POS from OSP
This is a stand-alone program to rebuild the last permutation pair from
an index build, which is PSO and POS, from the OSP permutation.

This is useful when the index build crashed during the creation of that
last permutation pair. The alternative would be to rebuild the entire
index from scratch, which is very time-consuming for large datasets.