Give more memory to twin-relation sorter for PSO/POS
So far, all twin-relation sorters were 4_GB of memory for merging. This
is fine for the SPO/SOP and OSP/OPS permutation pairs, where "relations"
(the number of triples with the same S or O) are usually small. However,
for PSO/POS, relations can be very large (think `rdf:type`), and 4_GB
can be too small for large datasets.

For that permutation pair, we now provide half of the memory available
for the whole index building (which is configurabel via --stxxl-memory).
That is OK because each permutation sorter gets half of that memory and
for the last permutation pair, no other pair is sorted at the same time.