Set proper cache key for `ExplicitIdTableOperation` (#2458)
The `NamedResultCache` from #1739 did not work as expected when using a named result from that cache in another query and then changing the named result in the cache. Then the query using the named result should be recomputed. This is now fixed, using the mechanism introduced in #2381, namely by setting the cache key of an `ExplicitIdTableOperation` to the cache key of the query used for obtaining the cached result.