Add `geof:numGeometries` using `GeometryInfo` (#2238)
Add the GeoSPARQL function `geof:numGeometries`, which counts the member geometries of a given literal. For example (the top 100 geometries with the most members):
```sparql
SELECT * WHERE {
  ?any geo:hasGeometry/geo:asWKT ?any_geometry .
  BIND (geof:numGeometries(?any_geometry) AS ?num)
}
ORDER BY DESC(?num)
LIMIT 100
```
NOTE: This is an index-breaking change with regard to the `on-disk-compressed-geo-split` vocabulary type. It will be merged in a sequence with #2172 and #2430