Speed up the precomputation for `geof:area` from #2430 (#2469)
The precomputation introduced in #2430 for `geof:area` and `geof:metricArea` took forever for the complete OSM data. This is now fixed by using `S2Polygon::DestructiveUnion` instead of `S2Polygon.InitToUnion` for computing the union of multipolygons. This avoids the quadratic running time when computing the covered area of possibly overlapping polygons. On the side, improve the logging regarding WKT literals for which the area cannot be computed.

NOTE: Even with this improvement, the running time of the vocabulary merging stage increases from 1.3 h (without area precomputation) to 2.6 h (with area precomputation) on the complete OSM data. There are over 1 B non-point geometries in that data and the area precomputation is currently single-threaded. A single area computation takes a few microseconds, which explains the 1.3 h increase in running time. The solution is to parallelize this.