Run expensive checks only if they are enabled [two instances] (#2501)
There were two expensive checks in `CompressedRelation.cpp` that were run for every index scan during query planning. One of them iterated over the complete block meta data, regardless of the query. This was particularly noticeable for large datasets, like `osm-planet`. The two tests are now both executed if and only if `areExpensiveChecksEnabled` is true, as it should have been in the first place.

NOTE: What still takes relatively long for large datasets, and now dominates the query planning time for simple queries, is `CompressedRelationReader::getResultSizeImpl`. We should look into this in a separate PR.