Avoid size estimate for full index scans in `TransitvePath`
When one side of a `TransitivePath` operation is constrained, the
eventuality of an "empty path" has to be taken into account. This is
achieved using two fulll index scans (SPO and OPS). The index scans are
efficient thanks to prefiltering, but the size estimates take
considerable time during query planning. However, the size estimates are
not actually used in this case. This change extends the `IndexScan`
constructor to add a predefined size estimate, which is used instead of
computing it. The `TransitivePath` class is modified to use this new
constructor when creating the two full index scans.