Show permutation of index scan in description and runtime information (#2504)
So far, the description of an index scan in the runtime information was like in `IndexScan ?x <p> ?z`. The information about the permutation was missing. Now the description has the form `IndexScan POS ?x <p> ?z`, where `POS` is the permutation of the index scan. In particular, this additional information now shows in the "Analysis" tree in the QLever UI. This is a very simple but useful change