Fix `GROUP BY` optimizations when operand has `LIMIT` or `OFFSET` (#2484)
Three of our `GROUP BY` optimizations (`computeGroupByForSingleIndexScan`, `computeGroupByForJoinWithFullScan`, and `computeGroupByObjectWithCount`) ignored the limit and offset of their operand. For example, the following query was executed as if the `OFFSET` and `LIMIT` weren't there.
```sparql
SELECT ?x (COUNT(?x) AS ?count) WHERE {
  { SELECT * { ?x ?y ?z } OFFSET 1 LIMIT 12 }
} GROUP BY ?x
```
This is now fixed by either correctly handling the limit and offset, or, when that is not possible, by disabling the optimization. Fixes #2465