Properly handle exceptions in `findUndefValuesHelper`
Since https://github.com/ad-freiburg/qlever/pull/2266, it could happen
that an exception is thrown in the destructor of `CallbackOnEndView`
(triggered by the end of the iteration over the range returned by
`findUndefValuesHelper`), which terminated the program.

This is now fixed by wrapping the call to `setInput` (which can throw)
in a `ThrowInDestructorIfSafe` object.