Reduce number of threads for small inputs to `libspatialjoinParse` (#2311)
So far, `SpatialJoinAlgorithms::libspatialjoinParse` always spawns exactly the number of threads as configured by the `spatial-join-max-num-threads` runtime parameter, which by default is 8. The parallel parser feeds the geometries to the threads in batches of 10,000. This change ensures that there are never more threads than batches. For example, if there are only 17,321 geometries, there will be two batches and only two threads (because the other threads would have nothing to do).