Respect UTF-8 boundaries in `truncateOperationString` (#2514)
At several places in our log and when returning query results, we write the string of the query or update operation. Especially for update operations, this string can be very long, so we truncate it to `MAX_LENGTH_OPERATION_ECHO`. So far, this truncation ignored boundaries of UTF-8 characters and thus potentially produced invalid UTF-8. This is now fixed. In particular, fixes #2511