Use `gtest` matchers for testing `GeometryInfo` and related helper classes (#2517)
This change migrates the test helpers for checking approximate equality of `GeoPoint`, `Centroid`, `BoundingBox`, `MetricLength`, `MetricArea` and `GeometryInfo` from custom helper functions to Gtest matchers. This is more in line with the intended way of using `gtest` and it allows, for example, storing a matcher in a variable and using it multiple times. Existing tests are rewritten as part of this change to use the new matchers instead of the previously used functions.

These changes are also a preparation for #2233, where this infrastructure will be extended for checking `GeoPointOrWkt` objects and a better infrastructure in `ValueGetterTestHelpers`.