Fast `FILTER` with `LANG` also for disjunction of languages (#2492)
So far, only language filter with a single condition, like `FILTER (LANG(?label) = "en")`, were rewritten using one of the special predicates, `@en@rdfs:label` in this case, which makes the query much faster. Now this trick is also applied for logical disjunctions like `FILTER (LANG(?label) = "de" || LANG(?label) = "en")` or `FILTER (LANG(?label) IN ("de", "en"))`. On the side, fix filtering by `LANG(...) = ""`,  which so far falsely returned an empty result because it translated the respective predicate to `@@...`, which never exists.

Inspired by #2363 (but the query that gave rise to the issue is still slow)