Reduce WebSocket communication for updates of runtime information (#2510)
QLever periodically sends the current runtime information over a WebSocket channel. For example, this is used by the QLever UI to show a live "Analysis" of the query while it is being processed. So far, each operation by itself sent the *complete* runtime information periodically. For complex queries, this both slowed down QLever (because it had to construct many large JSON strings with high frequency) and the QLever UI (because it had to consume them). This is now fixed by centralizing the check for how long ago the last update was in `QueryExecutionContext` (of which there is only one per query). Use the occasion to improve the code; in particular, there is now an `enum class SendPriority` with two values `IfDue` and `Always`.

NOTE: The JSON strings are still constructed periodically, regardless of whether they are ever consumed by anyone. This is work for a separate PR

Co-authored-by: Hannah Bast <bast@cs.uni-freiburg.de>