Faster vocabulary lookup for update operations (#2477)
A significant part of an `INSERT DATA { ... }` or `DELETE DATA { ... }` operation (or the corresponding Graph Store Protocol operation) is spent for looking up the IRIs and literals in the operation in QLever's vocabulary [NOTE 2]. So far, these lookups were made one after the other, in the order in which the IRIs or literals occurred in the operation, which typically translates into random lookup on disk (for the part of the vocabulary that resides there, usually the larger part) or in memory (for the part of the vocabulary that resides there). With this change, the lookups are lexicographically ordered (which significantly improves the locality of access) and deduplicated (which has the potential to save some lookups).

NOTE 1: To make the sorted lookups, each distinct `TripleComponent` from the update operation is currently copied multiple times. This does not seem to affect performance significantly though because the lookup time is dominated by reading the compressed strings from disk and decompressing them. The places were the copies are made are marked in the code.

NOTE 2: So far, an `INSERT DATA { ... }` operation constructed from 100,000 messages from the Wikidata update stream took around 60s with cold disk cache and around 40s with warm disk cache, with around 40s and 20s, respectively, spent on the vocabulary lookups. With this change, the operation time is reduced to around 30s with cold disk cache and around 20s with warm disk cache, with around 15s and 3s, respectively, spent on vocabulary lookups.

Co-authored-by: Hannah Bast <bast@cs.uni-freiburg.de>