Add `ql:has-word` triples to internal PSO&POS permutation
During parsing, for each triples with a literal object, and for each
word in that literal, add an internal triple `subject ql:has-word
"word"`.

TODO: This is currently done unconditionally, which makes it easier to
test (we don't need special options in the Qleverfile). Eventually,
there should be an option `--add-has-word-triples` to `IndexBuilderMain`
to enable this behavior. Tests are also still missing