Write term frequency, not position
Writing the position is more general. But computing the term frequencies
for each text-word pair is currently not efficient in QLever (it
requires too much memory and a GROUP BY with two variables is much
slower than a GROUP BY with one variable). Since we never needed
positions so far, but we do want term frequencies for scoring, let's
make this the default for now.