Enable allocation of an explicit set of ranges of blank node identifiers (#2574)
`CONSTRUCT` queries, `SERVICE` requests, update operations, and the `BNODE` function can create new blank nodes, which must be separate from the already existing blank nodes. To this end, #1504 introduced a `BlankNodeManager`, later extended and improved in #1617 and #2126. So far, there was no control over exactly which blank node identifiers where returned by that manager. However, this is important for persisting results containing such blank node identifiers across server restarts because multiple such results may refer to the same blank node identifiers.

There is now a mechanism for allocating an explicit set of ranges of blank node identifiers. Each such set is identified by a unique UUID, so that the same set of ranges can be unambiguously referenced by multiple results (we could have also used the ranges themselves as unique identifier, but a UUID is simpler). There is also infrastructure for integrating this information with the `LocalVocab` (queries and updates can generate both new blank nodes and new local vocab entries) and serializing this information. There is not yet code for actually writing it to disk or reading it back from disk, that is work for a separate PR