Minor fix of integration of internal permutations into `DeltaTriples` (#2590)
In #2561, the `locatedTriples_`, `triplesInserted_`, and `triplesDeleted_` became members of a new `struct`, which has one instance for the normal permutations and one instance for the internal permutations. However, for #2500, the `locatedTriples_` are needed separately from the other two. Therefore, they are now factored out again. On the side, rename the `struct` from the rather generic `State` to the more specific `TriplesToHandles`

Co-authored-by: Hannah Bast <bast@cs.uni-freiburg.de>